Crate pallet_subtensor

Re-exports

pub use pallet::*;

Modules

coinbase

epoch

macros

migrations

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

rpc_info

staking

subnets

swap

utils

Structs

SubtensorSignedExtension

Enums

CallType

CustomTransactionError

RateLimitKey

Enum that defines types of rate limited operations for storing last block when this operation occured

Constants

MAX_CRV3_COMMIT_SIZE_BYTES

Traits

CollectiveInterface

Trait for interacting with collective pallets

MemberManagement

Trait for managing a membership pallet instance in the runtime