pallet_subtensor

Module pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Re-exports

pub use __tt_error_token_1 as tt_error_token;

pub use __tt_default_parts_7 as tt_default_parts;

pub use __tt_extra_parts_7 as tt_extra_parts;

pub use __tt_default_parts_v2_7 as tt_default_parts_v2;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

AxonInfo

Data structure for Axon information.

ChainIdentity

Data structure for Chain Identities. (DEPRECATED for V2)

ChainIdentityV2

Data structure for Chain Identities.

DefaultAccount

Default account, derived from zero trailing bytes.

DefaultAccountLinkage

Default account linkage

DefaultAccountTake

Default account take.

DefaultAccumulatedEmission

Default accumulated emission for a hotkey

DefaultActivityCutoff

Default activity cutoff.

DefaultAdjustmentAlpha

Default adjustment alpha.

DefaultAdjustmentInterval

Default adjustment interval.

DefaultAllowsDelegation

Default allowed delegation.

DefaultAlphaSigmoidSteepness

Default value for alpha sigmoid steepness.

DefaultAlphaValues

(alpha_low: 0.7, alpha_high: 0.9)

DefaultBlockAtRegistration

Default block number at registration.

DefaultBlockEmission

Default emission per block.

DefaultBlocksSinceLastStep

Default value for blocks since last step.

DefaultBonds

Value definition for bonds with type vector of (u16, u16).

DefaultBondsMovingAverage

Default bonds moving average.

DefaultBondsPenalty

Default bonds penalty.

DefaultBondsResetOn

Default value for bonds reset - will not reset bonds

DefaultBurn

Default registrations this block.

DefaultChildKeyTake

Default childkey take.

DefaultColdkeySwapRescheduleDuration

Default value for coldkey swap reschedule duration

DefaultColdkeySwapScheduleDuration

Default value for coldkey swap schedule duration

DefaultColdkeySwapScheduled

Default value for coldkey swap scheduled

DefaultCommitRevealWeightsEnabled

Default value for weight commit/reveal enabled.

DefaultDelegateTake

Default Delegate Take.

DefaultDifficulty

Default max burn token.

DefaultDissolveNetworkScheduleDuration

Default value for dissolve network schedule duration

DefaultEMAPriceMovingBlocks

Default EMA price halving blocks

DefaultFalse

Default value for false.

DefaultHotkeys

Default value for hotkeys.

DefaultImmunityPeriod

Default immunity period.

DefaultIsNetworkMember

Default value for network member.

DefaultKappa

Default value for kappa parameter.

DefaultKey

Default value for key with type T::AccountId derived from trailing zeroes.

DefaultLastAdjustmentBlock

Default last adjustment block.

DefaultLastMechanismStepBlock

Default value for last mechanism step block.

DefaultLastTxBlock

Default value for last extrinsic block.

DefaultLiquidAlpha

– ITEM (switches liquid alpha on)

DefaultMaxAllowedUids

Default maximum allowed UIDs.

DefaultMaxAllowedValidators

Default maximum allowed validators.

DefaultMaxBurn

Default min burn token.

DefaultMaxChildKeyTake

Default maximum childkey take.

DefaultMaxDifficulty

Default min difficulty value.

DefaultMaxRegistrationsPerBlock

Default max difficulty value.

DefaultMaxWeightsLimit

Default maximum weights limit.

DefaultMinActivityCutoff

Default value for minimum activity cutoff

DefaultMinAllowedWeights

Default minimum allowed weights.

DefaultMinBurn

Default burn token.

DefaultMinChildKeyTake

Default minimum childkey take.

DefaultMinDelegateTake

Default minimum delegate take.

DefaultMinDifficulty

Default difficulty value.

DefaultMinStake

Default minimum stake.

DefaultMinimumPoolLiquidity

Default value for minimum liquidity in pool

DefaultModality

Default value for modality.

DefaultMovingAlpha

Default moving alpha for the moving price.

DefaultMovingPrice

Default subnet moving price.

DefaultN

Default number of networks.

DefaultNetworkImmunityPeriod

Default value for network immunity period.

DefaultNetworkLastRegistered

Default value for network last registered.

DefaultNetworkLockReductionInterval

Default value for network lock reduction interval.

DefaultNetworkMinAllowedUids

Default value for network min allowed UIDs.

DefaultNetworkMinLockCost

Default value for network min lock cost.

DefaultNetworkRateLimit

Default value for network rate limit.

DefaultNetworkRegisteredAt

Default value for network registered at.

DefaultNeworksAdded

Default value if network is added.

DefaultPendingChildkeys

Default pending childkeys

DefaultPendingCooldown

Default value for applying pending items (e.g. childkeys).

DefaultPendingEmission

Default value for pending emission.

DefaultProportion

Default account linkage

DefaultRAORecycledForRegistration

Default max registrations per block.

DefaultRegistrationAllowed

Default value for registration allowed.

DefaultRegistrationsThisBlock

Default last adjustment block.

DefaultRevealPeriodEpochs

Default Reveal Period Epochs

DefaultRho

Default value for rho parameter.

DefaultScalingLawPower

Default scaling law power.

DefaultSenateRequiredStakePercentage

Senate requirements

DefaultServingRateLimit

Default value for serving rate limit.

DefaultSetSNOwnerHotkeyRateLimit

Default value for setting subnet owner hotkey rate limit

DefaultSharePoolZero

Default value for Share Pool variables

DefaultStakeThreshold

Default minimum stake for weights.

DefaultStakingFee

Default staking fee.

DefaultSubnetLocked

Default value for subnet locked.

DefaultSubnetOwner

Default value for subnet owner.

DefaultSubnetOwnerCut

Default value for subnet owner cut.

DefaultTaoWeight

Default value for global weight.

DefaultTargetRegistrationsPerInterval

Default target registrations per interval.

DefaultTempo

Default value for network tempo

DefaultTotalIssuance

Default total issuance.

DefaultTrue

Default value for false.

DefaultTxChildKeyTakeRateLimit

Default value for chidlkey take rate limiting

DefaultTxDelegateTakeRateLimit

Default value for delegate take rate limiting

DefaultTxRateLimit

Default value for rate limiting

DefaultUnicodeVecU8

Default unicode vector for tau symbol.

DefaultValidatorPruneLen

Default validator prune length.

DefaultWeights

Value definition for weights with vector of (u16, u16).

DefaultWeightsSetRateLimit

Default value for weights set rate limit.

DefaultWeightsVersionKey

Default weights version key.

DefaultWeightsVersionKeyRateLimit

Default value for weights version key rate limit. In units of tempos.

DefaultYuma3

– ITEM (switches liquid alpha on)

DefaultZeroU16

Default value for zero.

DefaultZeroU64

============================ ==== Staking + Accounts ====

DefaultZeroU128

Default value for zero.

EmptyBoolVec

Value definition for vector of bool.

EmptyU16Vec

Value definition for vector of u16.

EmptyU64Vec

Value definition for vector of u64.

GenesisConfig

Can be used to configure the genesis state of this pallet.

NeuronCertificate

Data structure for NeuronCertificate information.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

PrometheusInfo

Data structure for Prometheus information.

SubnetIdentity

Data structure for Subnet Identities. (DEPRECATED for V2)

SubnetIdentityV2

Data structure for Subnet Identities

TotalSupply

Total Rao in circulation.

Enums

Call

Dispatchable functions allow users to interact with the pallet and invoke state changes. These functions materialize as “extrinsics”, which are often compared to transactions. Dispatchable functions must be annotated with a weight and must return a DispatchResult.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

StakeJob

Data structure for stake related jobs.

Constants

MIN_BALANCE_TO_PERFORM_COLDKEY_SWAP

Minimum balance required to perform a coldkey swap

Traits

Config

Configuration trait of this pallet.

Functions

__type_value_for_default_account

Default account, derived from zero trailing bytes.

__type_value_for_default_account_linkage

Default account linkage

__type_value_for_default_account_take

Default account take.

__type_value_for_default_accumulated_emission

Default accumulated emission for a hotkey

__type_value_for_default_activity_cutoff

Default activity cutoff.

__type_value_for_default_adjustment_alpha

Default adjustment alpha.

__type_value_for_default_adjustment_interval

Default adjustment interval.

__type_value_for_default_allows_delegation

Default allowed delegation.

__type_value_for_default_alpha_sigmoid_steepness

Default value for alpha sigmoid steepness.

__type_value_for_default_alpha_values

(alpha_low: 0.7, alpha_high: 0.9)

__type_value_for_default_block_at_registration

Default block number at registration.

__type_value_for_default_block_emission

Default emission per block.

__type_value_for_default_blocks_since_last_step

Default value for blocks since last step.

__type_value_for_default_bonds

Value definition for bonds with type vector of (u16, u16).

__type_value_for_default_bonds_moving_average

Default bonds moving average.

__type_value_for_default_bonds_penalty

Default bonds penalty.

__type_value_for_default_bonds_reset_on

Default value for bonds reset - will not reset bonds

__type_value_for_default_burn

Default registrations this block.

__type_value_for_default_child_key_take

Default childkey take.

__type_value_for_default_coldkey_swap_reschedule_duration

Default value for coldkey swap reschedule duration

__type_value_for_default_coldkey_swap_schedule_duration

Default value for coldkey swap schedule duration

__type_value_for_default_coldkey_swap_scheduled

Default value for coldkey swap scheduled

__type_value_for_default_commit_reveal_weights_enabled

Default value for weight commit/reveal enabled.

__type_value_for_default_delegate_take

Default Delegate Take.

__type_value_for_default_difficulty

Default max burn token.

__type_value_for_default_dissolve_network_schedule_duration

Default value for dissolve network schedule duration

__type_value_for_default_ema_price_moving_blocks

Default EMA price halving blocks

__type_value_for_default_false

Default value for false.

__type_value_for_default_hotkeys

Default value for hotkeys.

__type_value_for_default_immunity_period

Default immunity period.

__type_value_for_default_is_network_member

Default value for network member.

__type_value_for_default_kappa

Default value for kappa parameter.

__type_value_for_default_key

Default value for key with type T::AccountId derived from trailing zeroes.

__type_value_for_default_last_adjustment_block

Default last adjustment block.

__type_value_for_default_last_mechanism_step_block

Default value for last mechanism step block.

__type_value_for_default_last_tx_block

Default value for last extrinsic block.

__type_value_for_default_liquid_alpha

– ITEM (switches liquid alpha on)

__type_value_for_default_max_allowed_uids

Default maximum allowed UIDs.

__type_value_for_default_max_allowed_validators

Default maximum allowed validators.

__type_value_for_default_max_burn

Default min burn token.

__type_value_for_default_max_child_key_take

Default maximum childkey take.

__type_value_for_default_max_difficulty

Default min difficulty value.

__type_value_for_default_max_registrations_per_block

Default max difficulty value.

__type_value_for_default_max_weights_limit

Default maximum weights limit.

__type_value_for_default_min_activity_cutoff

Default value for minimum activity cutoff

__type_value_for_default_min_allowed_weights

Default minimum allowed weights.

__type_value_for_default_min_burn

Default burn token.

__type_value_for_default_min_child_key_take

Default minimum childkey take.

__type_value_for_default_min_delegate_take

Default minimum delegate take.

__type_value_for_default_min_difficulty

Default difficulty value.

__type_value_for_default_min_stake

Default minimum stake.

__type_value_for_default_minimum_pool_liquidity

Default value for minimum liquidity in pool

__type_value_for_default_modality

Default value for modality.

__type_value_for_default_moving_alpha

Default moving alpha for the moving price.

__type_value_for_default_moving_price

Default subnet moving price.

__type_value_for_default_n

Default number of networks.

__type_value_for_default_network_immunity_period

Default value for network immunity period.

__type_value_for_default_network_last_registered

Default value for network last registered.

__type_value_for_default_network_lock_reduction_interval

Default value for network lock reduction interval.

__type_value_for_default_network_min_allowed_uids

Default value for network min allowed UIDs.

__type_value_for_default_network_min_lock_cost

Default value for network min lock cost.

__type_value_for_default_network_rate_limit

Default value for network rate limit.

__type_value_for_default_network_registered_at

Default value for network registered at.

__type_value_for_default_neworks_added

Default value if network is added.

__type_value_for_default_pending_childkeys

Default pending childkeys

__type_value_for_default_pending_cooldown

Default value for applying pending items (e.g. childkeys).

__type_value_for_default_pending_emission

Default value for pending emission.

__type_value_for_default_proportion

Default account linkage

__type_value_for_default_rao_recycled_for_registration

Default max registrations per block.

__type_value_for_default_registration_allowed

Default value for registration allowed.

__type_value_for_default_registrations_this_block

Default last adjustment block.

__type_value_for_default_reveal_period_epochs

Default Reveal Period Epochs

__type_value_for_default_rho

Default value for rho parameter.

__type_value_for_default_scaling_law_power

Default scaling law power.

__type_value_for_default_senate_required_stake_percentage

Senate requirements

__type_value_for_default_serving_rate_limit

Default value for serving rate limit.

__type_value_for_default_set_sn_owner_hotkey_rate_limit

Default value for setting subnet owner hotkey rate limit

__type_value_for_default_share_pool_zero

Default value for Share Pool variables

__type_value_for_default_stake_threshold

Default minimum stake for weights.

__type_value_for_default_staking_fee

Default staking fee.

__type_value_for_default_subnet_locked

Default value for subnet locked.

__type_value_for_default_subnet_owner

Default value for subnet owner.

__type_value_for_default_subnet_owner_cut

Default value for subnet owner cut.

__type_value_for_default_tao_weight

Default value for global weight.

__type_value_for_default_target_registrations_per_interval

Default target registrations per interval.

__type_value_for_default_tempo

Default value for network tempo

__type_value_for_default_total_issuance

Default total issuance.

__type_value_for_default_true

Default value for false.

__type_value_for_default_tx_child_key_take_rate_limit

Default value for chidlkey take rate limiting

__type_value_for_default_tx_delegate_take_rate_limit

Default value for delegate take rate limiting

__type_value_for_default_tx_rate_limit

Default value for rate limiting

__type_value_for_default_unicode_vec_u8

Default unicode vector for tau symbol.

__type_value_for_default_validator_prune_len

Default validator prune length.

__type_value_for_default_weights

Value definition for weights with vector of (u16, u16).

__type_value_for_default_weights_set_rate_limit

Default value for weights set rate limit.

__type_value_for_default_weights_version_key

Default weights version key.

__type_value_for_default_weights_version_key_rate_limit

Default value for weights version key rate limit. In units of tempos.

__type_value_for_default_yuma_3

– ITEM (switches liquid alpha on)

__type_value_for_default_zero_u16

Default value for zero.

__type_value_for_default_zero_u64

============================ ==== Staking + Accounts ====

__type_value_for_default_zero_u128

Default value for zero.

__type_value_for_empty_bool_vec

Value definition for vector of bool.

__type_value_for_empty_u16_vec

Value definition for vector of u16.

__type_value_for_empty_u64_vec

Value definition for vector of u64.

__type_value_for_total_supply

Total Rao in circulation.

Type Aliases

AccountIdOf

Alias for the account ID.

Active

— MAP ( netuid ) –> active

ActivityCutoff

— MAP ( netuid ) –> activity_cutoff

AdjustmentAlpha

— MAP ( netuid ) –> adjustment_alpha

AdjustmentInterval

— MAP ( netuid ) –> adjustment_interval

Alpha

Storage type is StorageNMap with keys type (T :: AccountId, T :: AccountId, u16) and value type U64F64.

AlphaDividendsPerSubnet

Storage type is StorageDoubleMap with key1 type u16, key2 type T :: AccountId and value type u64.

AlphaSigmoidSteepness

— MAP ( netuid ) –> AlphaSigmoidSteepness

AlphaValues

MAP ( netuid ) –> (alpha_low, alpha_high)

AssociatedEvmAddress

============================= ==== EVM related storage ====

AxonInfoOf

Struct for Axon.

Axons

— MAP ( netuid, hotkey ) –> axon_info

BlockAtRegistration

— DMAP ( netuid, uid ) –> block_at_registration

BlockEmission

================== ==== Coinbase ====

BlocksSinceLastStep

— MAP ( netuid ) –> blocks_since_last_step

Bonds

— DMAP ( netuid, uid ) –> bonds

BondsMovingAverage

— MAP ( netuid ) –> bonds_moving_average

BondsPenalty

— MAP ( netuid ) –> bonds_penalty

BondsResetOn

— MAP ( netuid ) –> bonds_reset

Burn

— MAP ( netuid ) –> Burn

BurnRegistrationsThisInterval

— MAP ( netuid ) –> burn_registrations_this_interval

CRV3WeightCommits

— MAP (netuid, commit_epoch) –> VecDeque<(who, serialized_compressed_commit, reveal_round)> | Stores a queue of v3 commits for an account on a given netuid.

CallOf

Call type for the pallet

ChainIdentityOf

Struct for ChainIdentities. (DEPRECATED for V2)

ChainIdentityOfV2

Struct for ChainIdentities.

ChildKeys

DMAP ( parent, netuid ) –> Vec<(proportion,child)>

ChildkeyTake

DMAP ( hot, netuid ) –> take | Returns the hotkey childkey take for a specific subnet

ColdkeySwapRescheduleDuration

Storage type is StorageValue with value type BlockNumberFor < T >.

ColdkeySwapScheduleDuration

Storage type is StorageValue with value type BlockNumberFor < T >.

ColdkeySwapScheduled

Storage type is StorageMap with key type T :: AccountId and value type (BlockNumberFor < T > , T :: AccountId).

CommitRevealWeightsEnabled

— MAP ( netuid ) –> commit reveal v2 weights are enabled

Consensus

— MAP ( netuid ) –> consensus

Delegates

MAP ( hot ) –> take | Returns the hotkey delegation take. And signals that this key is open for delegation.

Difficulty

— MAP ( netuid ) –> Difficulty

DissolveNetworkScheduleDuration

Storage type is StorageValue with value type BlockNumberFor < T >.

Dividends

— MAP ( netuid ) –> dividends

EMAPriceHalvingBlocks

— MAP ( netuid ) –> Halving time of average moving price.

Emission

— MAP ( netuid ) –> emission

FirstEmissionBlockNumber

============================ ==== Subnet Parameters =====

HasMigrationRun

================== ==== Genesis =====

Identities

Storage type is StorageMap with key type T :: AccountId and value type ChainIdentityOf.

IdentitiesV2

Storage type is StorageMap with key type T :: AccountId and value type ChainIdentityOfV2.

ImmunityPeriod

— MAP ( netuid ) –> immunity_period

Incentive

— MAP ( netuid ) –> incentive

IsNetworkMember

— DMAP ( hotkey, netuid ) –> bool

Kappa

— MAP ( netuid ) –> Kappa

Keys

— DMAP ( netuid, uid ) –> hotkey

LargestLocked

Storage type is StorageMap with key type u16 and value type u64.

LastAdjustmentBlock

— MAP ( netuid ) –> Block at last adjustment.

LastColdkeyHotkeyStakeBlock

— Map (coldkey, hotkey) –> u64 the last block at which stake was added/removed.

LastHotkeyEmissionOnNetuid

— DMap ( hot, netuid ) –> emission | last hotkey emission on network.

LastMechansimStepBlock

— MAP ( netuid ) –> last_mechanism_step_block

LastRateLimitedBlock

============================ ==== Rate Limiting =====

LastTxBlock

— MAP ( key ) –> last_block

LastTxBlockChildKeyTake

— MAP ( key ) –> last_tx_block_childkey_take

LastTxBlockDelegateTake

— MAP ( key ) –> last_tx_block_delegate_take

LastUpdate

— MAP ( netuid ) –> last_update

LiquidAlphaOn

— MAP ( netuid ) –> Whether or not Liquid Alpha is enabled

LoadedEmission

— MAP ( netuid ) –> (hotkey, se, ve)

LocalCallOf

local one

MaxAllowedUids

— MAP ( netuid ) –> max_allowed_uids

MaxAllowedValidators

— MAP ( netuid ) –> max_allowed_validators

MaxBurn

— MAP ( netuid ) –> MaxBurn

MaxChildkeyTake

— ITEM ( default_childkey_take )

MaxDelegateTake

— ITEM ( default_delegate_take )

MaxDifficulty

— MAP ( netuid ) –> MaxDifficulty

MaxRegistrationsPerBlock

— ITEM( global_max_registrations_per_block )

MaxWeightsLimit

— MAP ( netuid ) –> max_weight_limit

MinActivityCutoff

Storage type is StorageValue with value type u16.

MinAllowedWeights

— MAP ( netuid ) –> min_allowed_weights

MinBurn

— MAP ( netuid ) –> MinBurn

MinChildkeyTake

— ITEM ( min_childkey_take )

MinDelegateTake

— ITEM ( min_delegate_take )

MinDifficulty

— MAP ( netuid ) –> MinDifficulty

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NetworkImmunityPeriod

ITEM( network_immunity_period )

NetworkLastLockCost

ITEM( last_network_lock_cost )

NetworkLastRegistered

ITEM( network_last_registered_block )

NetworkLockReductionInterval

ITEM( network_lock_reduction_interval )

NetworkMinLockCost

ITEM( min_network_lock_cost )

NetworkModality

— MAP ( netuid ) –> modality TEXT: 0, IMAGE: 1, TENSOR: 2

NetworkPowRegistrationAllowed

— MAP ( netuid ) –> network_pow_allowed

NetworkRateLimit

ITEM( network_rate_limit )

NetworkRegisteredAt

— MAP ( netuid ) –> block_created

NetworkRegistrationAllowed

— MAP ( netuid ) –> network_registration_allowed

NetworksAdded

— MAP ( netuid ) –> network_is_added

NeuronCertificateOf

Struct for NeuronCertificate.

NeuronCertificates

— MAP ( netuid, hotkey ) –> certificate

NextStakeJobId

Ensures unique IDs for StakeJobs storage map

NominatorMinRequiredStake

Storage type is StorageValue with value type u64.

OwnedHotkeys

Storage type is StorageMap with key type T :: AccountId and value type Vec < T :: AccountId >.

Owner

MAP ( hot ) –> cold | Returns the controlling coldkey for a hotkey.

POWRegistrationsThisInterval

— MAP ( netuid ) –> pow_registrations_this_interval

PalletsOriginOf

Origin for the pallet

ParentKeys

DMAP ( child, netuid ) –> Vec<(proportion,parent)>

PendingAlphaSwapped

— MAP ( netuid ) –> pending_alpha_swapped

PendingChildKeys

DMAP ( netuid, parent ) –> (Vec<(proportion,child)>, cool_down_block)

PendingEmission

— MAP ( netuid ) –> pending_emission

PendingOwnerCut

— MAP ( netuid ) –> pending_owner_cut

PendingRootDivs

— MAP ( netuid ) –> pending_root_emission

Prometheus

— MAP ( netuid, hotkey ) –> prometheus_info

PrometheusInfoOf

Struct for Prometheus.

PruningScores

— MAP ( netuid ) –> pruning_scores

RAORecycledForRegistration

— MAP ( netuid ) –> global_RAO_recycled_for_registration

Rank

— MAP ( netuid ) –> rank

RegistrationsThisBlock

— MAP ( netuid ) –> Registrations of this Block.

RegistrationsThisInterval

— MAP ( netuid ) –> registrations_this_interval

RevealPeriodEpochs

— Map (netuid) –> Number of epochs allowed for commit reveal periods

Rho

— MAP ( netuid ) –> Rho

ScalingLawPower

— MAP ( netuid ) –> scaling_law_power

SenateRequiredStakePercentage

Storage type is StorageValue with value type u64.

ServingRateLimit

— MAP ( netuid ) –> serving_rate_limit

StakeJobs

Storage type is StorageDoubleMap with key1 type BlockNumberFor < T >, key2 type u64 and value type StakeJob < T :: AccountId >.

StakeThreshold

ITEM( weights_min_stake )

StakingHotkeys

Storage type is StorageMap with key type T :: AccountId and value type Vec < T :: AccountId >.

SubnetAlphaIn

Storage type is StorageMap with key type u16 and value type u64.

SubnetAlphaInEmission

Storage type is StorageMap with key type u16 and value type u64.

SubnetAlphaOut

Storage type is StorageMap with key type u16 and value type u64.

SubnetAlphaOutEmission

Storage type is StorageMap with key type u16 and value type u64.

SubnetIdentities

Storage type is StorageMap with key type u16 and value type SubnetIdentityOf.

SubnetIdentitiesV2

Storage type is StorageMap with key type u16 and value type SubnetIdentityOfV2.

SubnetIdentityOf

Struct for SubnetIdentities. (DEPRECATED for V2)

SubnetIdentityOfV2

Struct for SubnetIdentitiesV2.

SubnetLocked

Storage type is StorageMap with key type u16 and value type u64.

SubnetMechanism

— MAP ( netuid ) –> subnet mechanism

SubnetMovingAlpha

Storage type is StorageValue with value type I96F32.

SubnetMovingPrice

Storage type is StorageMap with key type u16 and value type I96F32.

SubnetOwner

— MAP ( netuid ) –> subnet_owner

SubnetOwnerCut

ITEM( subnet_owner_cut )

SubnetOwnerHotkey

— MAP ( netuid ) –> subnet_owner_hotkey

SubnetTAO

Storage type is StorageMap with key type u16 and value type u64.

SubnetTaoInEmission

Storage type is StorageMap with key type u16 and value type u64.

SubnetVolume

Storage type is StorageMap with key type u16 and value type u128.

SubnetworkN

— MAP ( netuid ) –> subnetwork_n (Number of UIDs in the network).

SubtokenEnabled

— MAP ( netuid ) –> If subtoken trading enabled

TaoDividendsPerSubnet

Storage type is StorageDoubleMap with key1 type u16, key2 type T :: AccountId and value type u64.

TaoWeight

============================ ==== Staking Variables ====

TargetRegistrationsPerInterval

— MAP ( netuid ) –> target_registrations_this_interval

Tempo

================= ==== Tempos =====

TokenSymbol

Storage type is StorageMap with key type u16 and value type Vec < u8 >.

TotalHotkeyAlpha

Storage type is StorageDoubleMap with key1 type T :: AccountId, key2 type u16 and value type u64.

TotalHotkeyAlphaLastEpoch

Storage type is StorageDoubleMap with key1 type T :: AccountId, key2 type u16 and value type u64.

TotalHotkeyShares

DMAP ( hot, netuid ) –> total_alpha_shares | Returns the number of alpha shares for a hotkey on a subnet.

TotalIssuance

========================== ==== Staking Counters ====

TotalNetworks

— ITEM( total_number_of_existing_networks )

TotalStake

Storage type is StorageValue with value type u64.

TransactionKeyLastBlock

================================= ==== Axon / Promo Endpoints =====

TransferToggle

============================ ==== Subnet Locks =====

Trust

— MAP ( netuid ) –> trust

TxChildkeyTakeRateLimit

— ITEM ( tx_childkey_take_rate_limit )

TxDelegateTakeRateLimit

— ITEM ( tx_delegate_take_rate_limit )

TxRateLimit

— ITEM ( tx_rate_limit )

Uids

— DMAP ( netuid, hotkey ) –> uid

UsedWork

============================ ==== Global Parameters =====

ValidatorPermit

— MAP ( netuid ) –> validator_permit

ValidatorPruneLen

— MAP ( netuid ) –> validator_prune_len

ValidatorTrust

— MAP ( netuid ) –> validator_trust

WeightCommits

— MAP (netuid, who) –> VecDeque<(hash, commit_block, first_reveal_block, last_reveal_block)> | Stores a queue of commits for an account on a given netuid.

Weights

— DMAP ( netuid, uid ) –> weights

WeightsSetRateLimit

— MAP ( netuid ) –> weights_set_rate_limit

WeightsVersionKey

— MAP ( netuid ) –> weights_version_key

WeightsVersionKeyRateLimit

ITEM( weights_version_key_rate_limit ) — Rate limit in tempos.

Yuma3On

— MAP ( netuid ) –> Whether or not Yuma3 is enabled