pallet_subtensor::pallet

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

add_stake

— Adds stake to a hotkey. The call is made from the coldkey account linked in the hotkey. Only the associated coldkey is allowed to make staking and unstaking requests. This protects the neuron against attacks on its hotkey running in production code.

add_stake_limit

— Adds stake to a hotkey on a subnet with a price limit. This extrinsic allows to specify the limit price for alpha token at which or better (lower) the staking should execute.

adjust_senate

Attempt to adjust the senate membership to include a hotkey

associate_evm_key

Attempts to associate a hotkey with an EVM key.

batch_commit_weights

— Allows a hotkey to commit weight hashes for multiple netuids as a batch.

batch_reveal_weights

–– The implementation for batch revealing committed weights.

batch_set_weights

— Allows a hotkey to set weights for multiple netuids as a batch.

become_delegate

— Sets the key as a delegate.

burn_alpha

Burns alpha from a cold/hot key pair without reducing AlphaOut

burned_register

User register a new subnetwork via burning token

commit_crv3_weights

–– Used to commit encrypted commit-reveal v3 weight values to later be revealed.

commit_weights

–– Used to commit a hash of your weight values to later be revealed.

decrease_take

— Allows delegates to decrease its take value.

dissolve_network

Remove a user’s subnetwork The caller must be the owner of the network

faucet

Facility extrinsic for user to get taken from faucet It is only available when pow-faucet feature enabled Just deployed in testnet and devnet for testing purpose

increase_take

— Allows delegates to increase its take value. This call is rate-limited.

move_stake

–– The implementation for the extrinsic move_stake: Moves specified amount of stake from a hotkey to another across subnets.

recycle_alpha

Recycles alpha from a cold/hot key pair, reducing AlphaOut on a subnet

register

–– Registers a new neuron to the subnetwork.

register_network

User register a new subnetwork

register_network_with_identity

User register a new subnetwork

remove_stake

Remove stake from the staking account. The call must be made from the coldkey account attached to the neuron metadata. Only this key has permission to make staking and unstaking requests.

remove_stake_limit

— Removes stake from a hotkey on a subnet with a price limit. This extrinsic allows to specify the limit price for alpha token at which or better (higher) the staking should execute.

reveal_weights

–– Used to reveal the weights for a previously committed hash.

root_register

Register the hotkey to root network

schedule_dissolve_network

Schedule the dissolution of a network at a specified block number.

schedule_swap_coldkey

Schedules a coldkey swap operation to be executed at a future block.

serve_axon

Serves or updates axon /prometheus information for the neuron associated with the caller. If the caller is already registered the metadata is updated. If the caller is not registered this call throws NotRegistered.

serve_axon_tls

Same as serve_axon but takes a certificate as an extra optional argument. Serves or updates axon /prometheus information for the neuron associated with the caller. If the caller is already registered the metadata is updated. If the caller is not registered this call throws NotRegistered.

serve_prometheus

–– Set prometheus information for the neuron.

set_childkey_take

Sets the childkey take for a given hotkey.

set_children

Set a single child for a given hotkey on a specified network.

set_identity

–– Set prometheus information for the neuron.

set_subnet_identity

–– Set the identity information for a subnet.

set_tao_weights

Args:

set_weights

— Sets the caller weights for the incentive mechanism. The call can be made from the hotkey account so is potentially insecure, however, the damage of changing weights is minimal if caught early. This function includes all the checks that the passed weights meet the requirements. Stored as u16s they represent rational values in the range [0,1] which sum to 1 and can be interpreted as probabilities. The specific weights determine how inflation propagates outward from this peer.

start_call

Initiates a call on a subnet.

sudo

Authenticates a council proposal and dispatches a function call with Root origin.

sudo_set_max_childkey_take

Sets the maximum allowed childkey take.

sudo_set_min_childkey_take

Sets the minimum allowed childkey take.

sudo_set_tx_childkey_take_rate_limit

Sets the transaction rate limit for changing childkey take.

sudo_unchecked_weight

Authenticates a council proposal and dispatches a function call with Root origin. This function does not check the weight of the call, and instead allows the user to specify the weight of the call.

swap_coldkey

The extrinsic for user to change the coldkey associated with their account.

swap_hotkey

The extrinsic for user to change its hotkey

swap_stake

Swaps a specified amount of stake from one subnet to another, while keeping the same coldkey and hotkey.

swap_stake_limit

Swaps a specified amount of stake from one subnet to another, while keeping the same coldkey and hotkey.

transfer_stake

Transfers a specified amount of stake from one coldkey to another, optionally across subnets, while keeping the same hotkey.

try_associate_hotkey

Attempts to associate a hotkey with a coldkey.

unstake_all

–– The implementation for the extrinsic unstake_all: Removes all stake from a hotkey account across all subnets and adds it onto a coldkey.

unstake_all_alpha

–– The implementation for the extrinsic unstake_all: Removes all stake from a hotkey account across all subnets and adds it onto a coldkey.

vote

User vote on a proposal