pallet_subtensor::pallet::dispatchables

Function associate_evm_key

pub fn associate_evm_key<T: Config>(
    netuid: u16,
    hotkey: T::AccountId,
    evm_key: H160,
    block_number: u64,
    signature: Signature,
)

Attempts to associate a hotkey with an EVM key.

The signature will be checked to see if the recovered public key matches the evm_key provided.

The EVM key is expected to sign the message according to this formula to produce the signature:

keccak_256(hotkey ++ keccak_256(block_number))

Arguments

origin - The origin of the transaction, which must be signed by the coldkey that owns the hotkey.

netuid - The netuid that the hotkey belongs to.

hotkey - The hotkey associated with the origin.

evm_key - The EVM key to associate with the hotkey.

block_number - The block number used in the signature.

signature - A signed message by the evm_key containing the hotkey and the hashed block_number.

Errors

Returns an error if:

The transaction is not signed.

The hotkey is not owned by the origin coldkey.

The hotkey does not belong to the subnet identified by the netuid.

The EVM key cannot be recovered from the signature.

The EVM key recovered from the signature does not match the given EVM key.

Events

May emit a EvmKeyAssociated event on success

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::associate_evm_key.