pallet_subtensor::pallet::dispatchables

Function batch_set_weights

pub fn batch_set_weights<T: Config>(
    netuids: Vec<Compact<u16>>,
    weights: Vec<Vec<(Compact<u16>, Compact<u16>)>>,
    version_keys: Vec<Compact<u64>>,
)

— Allows a hotkey to set weights for multiple netuids as a batch.

Args:

origin: (Origin):

The caller, a hotkey who wishes to set their weights.

netuids (Vec<Compact>):

The network uids we are setting these weights on.

weights (Vec<Vec<(Compact, Compact)>):

The weights to set for each network. [(uid, weight), …]

version_keys (Vec<Compact>):

The network version keys to check if the validator is up to date.

Event:

WeightsSet;

On successfully setting the weights on chain.

BatchWeightsCompleted;

On success of the batch.

BatchCompletedWithErrors;

On failure of any of the weights in the batch.

BatchWeightItemFailed;

On failure for each failed item in the batch.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::batch_set_weights.