pallet_subtensor::pallet::dispatchables

Function commit_crv3_weights

pub fn commit_crv3_weights<T: Config>(
    netuid: u16,
    commit: BoundedVec<u8, ConstU32<MAX_CRV3_COMMIT_SIZE_BYTES>>,
    reveal_round: u64,
)

–– Used to commit encrypted commit-reveal v3 weight values to later be revealed.

Args:

origin: (<T as frame_system::Config>::RuntimeOrigin):

The committing hotkey.

netuid (u16):

The u16 network identifier.

commit (Vec<u8>):

The encrypted compressed commit. The steps for this are:

Instantiate [WeightsTlockPayload]

Serialize it using the parity_scale_codec::Encode trait

Encrypt it following the steps (here)[https://github.com/ideal-lab5/tle/blob/f8e6019f0fb02c380ebfa6b30efb61786dede07b/timelock/src/tlock.rs#L283-L336] to produce a [TLECiphertext<TinyBLS381>] type.

Serialize and compress using the ark-serialize CanonicalSerialize trait.

reveal_round (u64):

The drand reveal round which will be avaliable during epoch n+1 from the current epoch.

Raises:

CommitRevealV3Disabled:

Attempting to commit when the commit-reveal mechanism is disabled.

TooManyUnrevealedCommits:

Attempting to commit when the user has more than the allowed limit of unrevealed commits.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::commit_crv3_weights.