pallet_subtensor::pallet::dispatchables

Function reveal_weights

pub fn reveal_weights<T: Config>(
    netuid: u16,
    uids: Vec<u16>,
    values: Vec<u16>,
    salt: Vec<u16>,
    version_key: u64,
)

–– Used to reveal the weights for a previously committed hash.

Args:

origin: (<T as frame_system::Config>::RuntimeOrigin):

The signature of the revealing hotkey.

netuid (u16):

The u16 network identifier.

uids (Vec<u16>):

The uids for the weights being revealed.

values (Vec<u16>):

The values of the weights being revealed.

salt (Vec<u16>):

The salt used to generate the commit hash.

version_key (u64):

The network version key.

Raises:

CommitRevealDisabled:

Attempting to reveal weights when the commit-reveal mechanism is disabled.

NoWeightsCommitFound:

Attempting to reveal weights without an existing commit.

ExpiredWeightCommit:

Attempting to reveal a weight commit that has expired.

RevealTooEarly:

Attempting to reveal weights outside the valid reveal period.

InvalidRevealCommitHashNotMatch:

The revealed hash does not match any committed hash.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::reveal_weights.