pallet_subtensor::pallet::dispatchables

Function schedule_swap_coldkey

pub fn schedule_swap_coldkey<T: Config>(new_coldkey: T::AccountId)

Schedules a coldkey swap operation to be executed at a future block.

This function allows a user to schedule the swapping of their coldkey to a new one at a specified future block. The swap is not executed immediately but is scheduled to occur at the specified block number.

Arguments

origin - The origin of the call, which should be signed by the current coldkey owner.

new_coldkey - The account ID of the new coldkey that will replace the current one.

when - The block number at which the coldkey swap should be executed.

Returns

Returns a DispatchResultWithPostInfo indicating whether the scheduling was successful.

Errors

This function may return an error if:

The origin is not signed.

The scheduling fails due to conflicts or system constraints.

Notes

The actual swap is not performed by this function. It merely schedules the swap operation.

The weight of this call is set to a fixed value and may need adjustment based on benchmarking.

TODO

Implement proper weight calculation based on the complexity of the operation.

Consider adding checks to prevent scheduling too far into the future. TODO: Benchmark this call

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::schedule_swap_coldkey.