pallet_subtensor::pallet::dispatchables

Function set_childkey_take

pub fn set_childkey_take<T: Config>(
    hotkey: T::AccountId,
    netuid: u16,
    take: u16,
)

Sets the childkey take for a given hotkey.

This function allows a coldkey to set the childkey take for a given hotkey. The childkey take determines the proportion of stake that the hotkey keeps for itself when distributing stake to its children.

Arguments:

origin (::RuntimeOrigin):

The signature of the calling coldkey. Setting childkey take can only be done by the coldkey.

hotkey (T::AccountId):

The hotkey for which the childkey take will be set.

take (u16):

The new childkey take value. This is a percentage represented as a value between 0 and 10000, where 10000 represents 100%.

Events:

ChildkeyTakeSet:

On successfully setting the childkey take for a hotkey.

Errors:

NonAssociatedColdKey:

The coldkey does not own the hotkey.

InvalidChildkeyTake:

The provided take value is invalid (greater than the maximum allowed take).

TxChildkeyTakeRateLimitExceeded:

The rate limit for changing childkey take has been exceeded.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::set_childkey_take.