pallet_subtensor::pallet::dispatchables

Function set_children

pub fn set_children<T: Config>(
    hotkey: T::AccountId,
    netuid: u16,
    children: Vec<(u64, T::AccountId)>,
)

Set a single child for a given hotkey on a specified network.

This function allows a coldkey to set a single child for a given hotkey on a specified network. The proportion of the hotkey’s stake to be allocated to the child is also specified.

Arguments:

origin (::RuntimeOrigin):

The signature of the calling coldkey. Setting a hotkey child can only be done by the coldkey.

hotkey (T::AccountId):

The hotkey which will be assigned the child.

child (T::AccountId):

The child which will be assigned to the hotkey.

netuid (u16):

The u16 network identifier where the childkey will exist.

proportion (u64):

Proportion of the hotkey’s stake to be given to the child, the value must be u64 normalized.

Events:

ChildAddedSingular:

On successfully registering a child to a hotkey.

Errors:

SubNetworkDoesNotExist:

Attempting to register to a non-existent network.

RegistrationNotPermittedOnRootSubnet:

Attempting to register a child on the root network.

NonAssociatedColdKey:

The coldkey does not own the hotkey or the child is the same as the hotkey.

HotKeyAccountNotExists:

The hotkey account does not exist.

Detailed Explanation of Checks:

Signature Verification: Ensures that the caller has signed the transaction, verifying the coldkey.

Root Network Check: Ensures that the delegation is not on the root network, as child hotkeys are not valid on the root.

Network Existence Check: Ensures that the specified network exists.

Ownership Verification: Ensures that the coldkey owns the hotkey.

Hotkey Account Existence Check: Ensures that the hotkey account already exists.

Child-Hotkey Distinction: Ensures that the child is not the same as the hotkey.

Old Children Cleanup: Removes the hotkey from the parent list of its old children.

New Children Assignment: Assigns the new child to the hotkey and updates the parent list for the new child.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::set_children.