pallet_subtensor::pallet::dispatchables

Function sudo_set_min_childkey_take

pub fn sudo_set_min_childkey_take<T: Config>(take: u16)

Sets the minimum allowed childkey take.

This function can only be called by the root origin.

Arguments:

origin - The origin of the call, must be root.

take - The new minimum childkey take value.

Errors:

BadOrigin - If the origin is not root.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::sudo_set_min_childkey_take.