pallet_subtensor::pallet::dispatchables

Function swap_stake_limit

pub fn swap_stake_limit<T: Config>(
    hotkey: T::AccountId,
    origin_netuid: u16,
    destination_netuid: u16,
    alpha_amount: u64,
    limit_price: u64,
    allow_partial: bool,
)

Swaps a specified amount of stake from one subnet to another, while keeping the same coldkey and hotkey.

Arguments

origin - The origin of the transaction, which must be signed by the coldkey that owns the hotkey.

hotkey - The hotkey whose stake is being swapped.

origin_netuid - The network/subnet ID from which stake is removed.

destination_netuid - The network/subnet ID to which stake is added.

alpha_amount - The amount of stake to swap.

limit_price - The limit price expressed in units of RAO per one Alpha.

allow_partial - Allows partial execution of the amount. If set to false, this becomes fill or kill type or order.

Errors

Returns an error if:

The transaction is not signed by the correct coldkey (i.e., coldkey_owns_hotkey fails).

Either origin_netuid or destination_netuid does not exist.

The hotkey does not exist.

There is insufficient stake on (coldkey, hotkey, origin_netuid).

The swap amount is below the minimum stake requirement.

Events

May emit a StakeSwapped event on success.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::swap_stake_limit.