pallet_subtensor::pallet::dispatchables

Function unstake_all

pub fn unstake_all<T: Config>(hotkey: T::AccountId)

–– The implementation for the extrinsic unstake_all: Removes all stake from a hotkey account across all subnets and adds it onto a coldkey.

Args:

origin - (::Origin):

The signature of the caller’s coldkey.

hotkey (T::AccountId):

The associated hotkey account.

Event:

StakeRemoved;

On the successfully removing stake from the hotkey account.

Raises:

NotRegistered:

Thrown if the account we are attempting to unstake from is non existent.

NonAssociatedColdKey:

Thrown if the coldkey does not own the hotkey we are unstaking from.

NotEnoughStakeToWithdraw:

Thrown if there is not enough stake on the hotkey to withdraw this amount.

TxRateLimitExceeded:

Thrown if key has hit transaction rate limit

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::unstake_all.