pallet_subtensor::pallet

Enum Call

pub enum Call<T: Config> {

    set_weights {
        netuid: u16,
        dests: Vec<u16>,
        weights: Vec<u16>,
        version_key: u64,
    },
    batch_set_weights {
        netuids: Vec<Compact<u16>>,
        weights: Vec<Vec<(Compact<u16>, Compact<u16>)>>,
        version_keys: Vec<Compact<u64>>,
    },
    commit_weights {
        netuid: u16,
        commit_hash: H256,
    },
    batch_commit_weights {
        netuids: Vec<Compact<u16>>,
        commit_hashes: Vec<H256>,
    },
    reveal_weights {
        netuid: u16,
        uids: Vec<u16>,
        values: Vec<u16>,
        salt: Vec<u16>,
        version_key: u64,
    },
    commit_crv3_weights {
        netuid: u16,
        commit: BoundedVec<u8, ConstU32<MAX_CRV3_COMMIT_SIZE_BYTES>>,
        reveal_round: u64,
    },
    batch_reveal_weights {
        netuid: u16,
        uids_list: Vec<Vec<u16>>,
        values_list: Vec<Vec<u16>>,
        salts_list: Vec<Vec<u16>>,
        version_keys: Vec<u64>,
    },
    set_tao_weights {
        netuid: u16,
        hotkey: T::AccountId,
        dests: Vec<u16>,
        weights: Vec<u16>,
        version_key: u64,
    },
    become_delegate {
        hotkey: T::AccountId,
    },
    decrease_take {
        hotkey: T::AccountId,
        take: u16,
    },
    increase_take {
        hotkey: T::AccountId,
        take: u16,
    },
    add_stake {
        hotkey: T::AccountId,
        netuid: u16,
        amount_staked: u64,
    },
    remove_stake {
        hotkey: T::AccountId,
        netuid: u16,
        amount_unstaked: u64,
    },
    serve_axon {
        netuid: u16,
        version: u32,
        ip: u128,
        port: u16,
        ip_type: u8,
        protocol: u8,
        placeholder1: u8,
        placeholder2: u8,
    },
    serve_axon_tls {
        netuid: u16,
        version: u32,
        ip: u128,
        port: u16,
        ip_type: u8,
        protocol: u8,
        placeholder1: u8,
        placeholder2: u8,
        certificate: Vec<u8>,
    },
    serve_prometheus {
        netuid: u16,
        version: u32,
        ip: u128,
        port: u16,
        ip_type: u8,
    },
    register {
        netuid: u16,
        block_number: u64,
        nonce: u64,
        work: Vec<u8>,
        hotkey: T::AccountId,
        coldkey: T::AccountId,
    },
    root_register {
        hotkey: T::AccountId,
    },
    adjust_senate {
        hotkey: T::AccountId,
    },
    burned_register {
        netuid: u16,
        hotkey: T::AccountId,
    },
    swap_hotkey {
        hotkey: T::AccountId,
        new_hotkey: T::AccountId,
    },
    swap_coldkey {
        old_coldkey: T::AccountId,
        new_coldkey: T::AccountId,
        swap_cost: u64,
    },
    set_childkey_take {
        hotkey: T::AccountId,
        netuid: u16,
        take: u16,
    },
    sudo_set_tx_childkey_take_rate_limit {
        tx_rate_limit: u64,
    },
    sudo_set_min_childkey_take {
        take: u16,
    },
    sudo_set_max_childkey_take {
        take: u16,
    },
    sudo {
        call: Box<T::SudoRuntimeCall>,
    },
    sudo_unchecked_weight {
        call: Box<T::SudoRuntimeCall>,
        weight: Weight,
    },
    vote {
        hotkey: T::AccountId,
        proposal: T::Hash,
        index: u32,
        approve: bool,
    },
    register_network {
        hotkey: T::AccountId,
    },
    faucet {
        block_number: u64,
        nonce: u64,
        work: Vec<u8>,
    },
    dissolve_network {
        coldkey: T::AccountId,
        netuid: u16,
    },
    set_children {
        hotkey: T::AccountId,
        netuid: u16,
        children: Vec<(u64, T::AccountId)>,
    },
    schedule_swap_coldkey {
        new_coldkey: T::AccountId,
    },
    schedule_dissolve_network {
        netuid: u16,
    },
    set_identity {
        name: Vec<u8>,
        url: Vec<u8>,
        github_repo: Vec<u8>,
        image: Vec<u8>,
        discord: Vec<u8>,
        description: Vec<u8>,
        additional: Vec<u8>,
    },
    set_subnet_identity {
        netuid: u16,
        subnet_name: Vec<u8>,
        github_repo: Vec<u8>,
        subnet_contact: Vec<u8>,
        subnet_url: Vec<u8>,
        discord: Vec<u8>,
        description: Vec<u8>,
        additional: Vec<u8>,
    },
    register_network_with_identity {
        hotkey: T::AccountId,
        identity: Option<SubnetIdentityOfV2>,
    },
    unstake_all {
        hotkey: T::AccountId,
    },
    unstake_all_alpha {
        hotkey: T::AccountId,
    },
    move_stake {
        origin_hotkey: T::AccountId,
        destination_hotkey: T::AccountId,
        origin_netuid: u16,
        destination_netuid: u16,
        alpha_amount: u64,
    },
    transfer_stake {
        destination_coldkey: T::AccountId,
        hotkey: T::AccountId,
        origin_netuid: u16,
        destination_netuid: u16,
        alpha_amount: u64,
    },
    swap_stake {
        hotkey: T::AccountId,
        origin_netuid: u16,
        destination_netuid: u16,
        alpha_amount: u64,
    },
    add_stake_limit {
        hotkey: T::AccountId,
        netuid: u16,
        amount_staked: u64,
        limit_price: u64,
        allow_partial: bool,
    },
    remove_stake_limit {
        hotkey: T::AccountId,
        netuid: u16,
        amount_unstaked: u64,
        limit_price: u64,
        allow_partial: bool,
    },
    swap_stake_limit {
        hotkey: T::AccountId,
        origin_netuid: u16,
        destination_netuid: u16,
        alpha_amount: u64,
        limit_price: u64,
        allow_partial: bool,
    },
    try_associate_hotkey {
        hotkey: T::AccountId,
    },
    start_call {
        netuid: u16,
    },
    associate_evm_key {
        netuid: u16,
        hotkey: T::AccountId,
        evm_key: H160,
        block_number: u64,
        signature: Signature,
    },
    recycle_alpha {
        hotkey: T::AccountId,
        amount: u64,
        netuid: u16,
    },
    burn_alpha {
        hotkey: T::AccountId,
        amount: u64,
        netuid: u16,
    },
    // some variants omitted

}

Dispatchable functions allow users to interact with the pallet and invoke state changes. These functions materialize as “extrinsics”, which are often compared to transactions. Dispatchable functions must be annotated with a weight and must return a DispatchResult.

Variants

set_weights

— Sets the caller weights for the incentive mechanism. The call can be made from the hotkey account so is potentially insecure, however, the damage of changing weights is minimal if caught early. This function includes all the checks that the passed weights meet the requirements. Stored as u16s they represent rational values in the range [0,1] which sum to 1 and can be interpreted as probabilities. The specific weights determine how inflation propagates outward from this peer.

Note: The 16 bit integers weights should represent 1.0 as the max u16. However, the function normalizes all integers to u16_max anyway. This means that if the sum of all elements is larger or smaller than the amount of elements * u16_max, all elements will be corrected for this deviation.

Args:

origin: (Origin):

The caller, a hotkey who wishes to set their weights.

netuid (u16):

The network uid we are setting these weights on.

dests (Vec):

The edge endpoint for the weight, i.e. j for w_ij.

‘weights’ (Vec):

The u16 integer encoded weights. Interpreted as rational values in the range [0,1]. They must sum to in32::MAX.

‘version_key’ ( u64 ):

The network version key to check if the validator is up to date.

Event:

WeightsSet;

On successfully setting the weights on chain.

Raises:

‘SubNetworkDoesNotExist’:

Attempting to set weights on a non-existent network.

‘NotRegistered’:

Attempting to set weights from a non registered account.

‘WeightVecNotEqualSize’:

Attempting to set weights with uids not of same length.

‘DuplicateUids’:

Attempting to set weights with duplicate uids.

‘UidsLengthExceedUidsInSubNet’:

Attempting to set weights above the max allowed uids.

‘UidVecContainInvalidOne’:

Attempting to set weights with invalid uids.

‘WeightVecLengthIsLow’:

Attempting to set weights with fewer weights than min.

‘MaxWeightExceeded’:

Attempting to set weights with max value exceeding limit.

Fields

netuid: u16

dests: Vec<u16>

weights: Vec<u16>

version_key: u64

batch_set_weights

— Allows a hotkey to set weights for multiple netuids as a batch.

Args:

origin: (Origin):

The caller, a hotkey who wishes to set their weights.

netuids (Vec<Compact>):

The network uids we are setting these weights on.

weights (Vec<Vec<(Compact, Compact)>):

The weights to set for each network. [(uid, weight), …]

version_keys (Vec<Compact>):

The network version keys to check if the validator is up to date.

Event:

WeightsSet;

On successfully setting the weights on chain.

BatchWeightsCompleted;

On success of the batch.

BatchCompletedWithErrors;

On failure of any of the weights in the batch.

BatchWeightItemFailed;

On failure for each failed item in the batch.

Fields

netuids: Vec<Compact<u16>>

weights: Vec<Vec<(Compact<u16>, Compact<u16>)>>

version_keys: Vec<Compact<u64>>

commit_weights

–– Used to commit a hash of your weight values to later be revealed.

Args:

origin: (<T as frame_system::Config>::RuntimeOrigin):

The signature of the committing hotkey.

netuid (u16):

The u16 network identifier.

commit_hash (H256):

The hash representing the committed weights.

Raises:

CommitRevealDisabled:

Attempting to commit when the commit-reveal mechanism is disabled.

TooManyUnrevealedCommits:

Attempting to commit when the user has more than the allowed limit of unrevealed commits.

Fields

netuid: u16

commit_hash: H256

batch_commit_weights

— Allows a hotkey to commit weight hashes for multiple netuids as a batch.

Args:

origin: (Origin):

The caller, a hotkey who wishes to set their weights.

netuids (Vec<Compact>):

The network uids we are setting these weights on.

commit_hashes (Vec):

The commit hashes to commit.

Event:

WeightsSet;

On successfully setting the weights on chain.

BatchWeightsCompleted;

On success of the batch.

BatchCompletedWithErrors;

On failure of any of the weights in the batch.

BatchWeightItemFailed;

On failure for each failed item in the batch.

Fields

netuids: Vec<Compact<u16>>

commit_hashes: Vec<H256>

reveal_weights

–– Used to reveal the weights for a previously committed hash.

Args:

origin: (<T as frame_system::Config>::RuntimeOrigin):

The signature of the revealing hotkey.

netuid (u16):

The u16 network identifier.

uids (Vec<u16>):

The uids for the weights being revealed.

values (Vec<u16>):

The values of the weights being revealed.

salt (Vec<u16>):

The salt used to generate the commit hash.

version_key (u64):

The network version key.

Raises:

CommitRevealDisabled:

Attempting to reveal weights when the commit-reveal mechanism is disabled.

NoWeightsCommitFound:

Attempting to reveal weights without an existing commit.

ExpiredWeightCommit:

Attempting to reveal a weight commit that has expired.

RevealTooEarly:

Attempting to reveal weights outside the valid reveal period.

InvalidRevealCommitHashNotMatch:

The revealed hash does not match any committed hash.

Fields

netuid: u16

uids: Vec<u16>

values: Vec<u16>

salt: Vec<u16>

version_key: u64

commit_crv3_weights

–– Used to commit encrypted commit-reveal v3 weight values to later be revealed.

Args:

origin: (<T as frame_system::Config>::RuntimeOrigin):

The committing hotkey.

netuid (u16):

The u16 network identifier.

commit (Vec<u8>):

The encrypted compressed commit. The steps for this are:

Instantiate [WeightsTlockPayload]

Serialize it using the parity_scale_codec::Encode trait

Encrypt it following the steps (here)[https://github.com/ideal-lab5/tle/blob/f8e6019f0fb02c380ebfa6b30efb61786dede07b/timelock/src/tlock.rs#L283-L336] to produce a [TLECiphertext<TinyBLS381>] type.

Serialize and compress using the ark-serialize CanonicalSerialize trait.

reveal_round (u64):

The drand reveal round which will be avaliable during epoch n+1 from the current epoch.

Raises:

CommitRevealV3Disabled:

Attempting to commit when the commit-reveal mechanism is disabled.

TooManyUnrevealedCommits:

Attempting to commit when the user has more than the allowed limit of unrevealed commits.

Fields

netuid: u16

commit: BoundedVec<u8, ConstU32<MAX_CRV3_COMMIT_SIZE_BYTES>>

reveal_round: u64

batch_reveal_weights

–– The implementation for batch revealing committed weights.

Args:

origin: (<T as frame_system::Config>::RuntimeOrigin):

The signature of the revealing hotkey.

netuid (u16):

The u16 network identifier.

uids_list (Vec<Vec<u16>>):

A list of uids for each set of weights being revealed.

values_list (Vec<Vec<u16>>):

A list of values for each set of weights being revealed.

salts_list (Vec<Vec<u16>>):

A list of salts used to generate the commit hashes.

version_keys (Vec<u64>):

A list of network version keys.

Raises:

CommitRevealDisabled:

Attempting to reveal weights when the commit-reveal mechanism is disabled.

NoWeightsCommitFound:

Attempting to reveal weights without an existing commit.

ExpiredWeightCommit:

Attempting to reveal a weight commit that has expired.

RevealTooEarly:

Attempting to reveal weights outside the valid reveal period.

InvalidRevealCommitHashNotMatch:

The revealed hash does not match any committed hash.

InvalidInputLengths:

The input vectors are of mismatched lengths.

Fields

netuid: u16

uids_list: Vec<Vec<u16>>

values_list: Vec<Vec<u16>>

salts_list: Vec<Vec<u16>>

version_keys: Vec<u64>

set_tao_weights

Args:

origin: (Origin):

The caller, a hotkey who wishes to set their weights.

netuid (u16):

The network uid we are setting these weights on.

hotkey (T::AccountId):

The hotkey associated with the operation and the calling coldkey.

dests (Vec):

The edge endpoint for the weight, i.e. j for w_ij.

‘weights’ (Vec):

The u16 integer encoded weights. Interpreted as rational values in the range [0,1]. They must sum to in32::MAX.

‘version_key’ ( u64 ):

The network version key to check if the validator is up to date.

Event:

WeightsSet;

On successfully setting the weights on chain.

Raises:

NonAssociatedColdKey;

Attempting to set weights on a non-associated cold key.

‘SubNetworkDoesNotExist’:

Attempting to set weights on a non-existent network.

‘NotRootSubnet’:

Attempting to set weights on a subnet that is not the root network.

‘WeightVecNotEqualSize’:

Attempting to set weights with uids not of same length.

‘UidVecContainInvalidOne’:

Attempting to set weights with invalid uids.

‘NotRegistered’:

Attempting to set weights from a non registered account.

‘WeightVecLengthIsLow’:

Attempting to set weights with fewer weights than min.

‘IncorrectWeightVersionKey’:

Attempting to set weights with the incorrect network version key.

‘SettingWeightsTooFast’:

Attempting to set weights too fast.

‘WeightVecLengthIsLow’:

Attempting to set weights with fewer weights than min.

‘MaxWeightExceeded’:

Attempting to set weights with max value exceeding limit.

Fields

netuid: u16

hotkey: T::AccountId

dests: Vec<u16>

weights: Vec<u16>

version_key: u64

become_delegate

— Sets the key as a delegate.

Args:

‘origin’: (Origin):

The signature of the caller’s coldkey.

‘hotkey’ (T::AccountId):

The hotkey we are delegating (must be owned by the coldkey.)

‘take’ (u64):

The stake proportion that this hotkey takes from delegations.

Event:

DelegateAdded;

On successfully setting a hotkey as a delegate.

Raises:

‘NotRegistered’:

The hotkey we are delegating is not registered on the network.

‘NonAssociatedColdKey’:

The hotkey we are delegating is not owned by the calling coldket.

Fields

hotkey: T::AccountId

decrease_take

— Allows delegates to decrease its take value.

Args:

‘origin’: (::Origin):

The signature of the caller’s coldkey.

‘hotkey’ (T::AccountId):

The hotkey we are delegating (must be owned by the coldkey.)

‘netuid’ (u16):

Subnet ID to decrease take for

‘take’ (u16):

The new stake proportion that this hotkey takes from delegations. The new value can be between 0 and 11_796 and should be strictly lower than the previous value. It T is the new value (rational number), the the parameter is calculated as [65535 * T]. For example, 1% would be [0.01 * 65535] = [655.35] = 655

Event:

TakeDecreased;

On successfully setting a decreased take for this hotkey.

Raises:

‘NotRegistered’:

The hotkey we are delegating is not registered on the network.

‘NonAssociatedColdKey’:

The hotkey we are delegating is not owned by the calling coldkey.

‘DelegateTakeTooLow’:

The delegate is setting a take which is not lower than the previous.

Fields

hotkey: T::AccountId

take: u16

increase_take

— Allows delegates to increase its take value. This call is rate-limited.

Args:

‘origin’: (::Origin):

The signature of the caller’s coldkey.

‘hotkey’ (T::AccountId):

The hotkey we are delegating (must be owned by the coldkey.)

‘take’ (u16):

The new stake proportion that this hotkey takes from delegations. The new value can be between 0 and 11_796 and should be strictly greater than the previous value. T is the new value (rational number), the the parameter is calculated as [65535 * T]. For example, 1% would be [0.01 * 65535] = [655.35] = 655

Event:

TakeIncreased;

On successfully setting a increased take for this hotkey.

Raises:

‘NotRegistered’:

The hotkey we are delegating is not registered on the network.

‘NonAssociatedColdKey’:

The hotkey we are delegating is not owned by the calling coldkey.

‘DelegateTakeTooHigh’:

The delegate is setting a take which is not greater than the previous.

Fields

hotkey: T::AccountId

take: u16

add_stake

— Adds stake to a hotkey. The call is made from the coldkey account linked in the hotkey. Only the associated coldkey is allowed to make staking and unstaking requests. This protects the neuron against attacks on its hotkey running in production code.

Args:

‘origin’: (Origin):

The signature of the caller’s coldkey.

‘hotkey’ (T::AccountId):

The associated hotkey account.

‘netuid’ (u16):

Subnetwork UID

‘amount_staked’ (u64):

The amount of stake to be added to the hotkey staking account.

Event:

StakeAdded;

On the successfully adding stake to a global account.

Raises:

‘NotEnoughBalanceToStake’:

Not enough balance on the coldkey to add onto the global account.

‘NonAssociatedColdKey’:

The calling coldkey is not associated with this hotkey.

‘BalanceWithdrawalError’:

Errors stemming from transaction pallet.

Fields

hotkey: T::AccountId

netuid: u16

amount_staked: u64

remove_stake

Remove stake from the staking account. The call must be made from the coldkey account attached to the neuron metadata. Only this key has permission to make staking and unstaking requests.

Args:

‘origin’: (Origin):

The signature of the caller’s coldkey.

‘hotkey’ (T::AccountId):

The associated hotkey account.

‘netuid’ (u16):

Subnetwork UID

‘amount_unstaked’ (u64):

The amount of stake to be added to the hotkey staking account.

Event:

StakeRemoved;

On the successfully removing stake from the hotkey account.

Raises:

‘NotRegistered’:

Thrown if the account we are attempting to unstake from is non existent.

‘NonAssociatedColdKey’:

Thrown if the coldkey does not own the hotkey we are unstaking from.

‘NotEnoughStakeToWithdraw’:

Thrown if there is not enough stake on the hotkey to withdwraw this amount.

Fields

hotkey: T::AccountId

netuid: u16

amount_unstaked: u64

serve_axon

Serves or updates axon /prometheus information for the neuron associated with the caller. If the caller is already registered the metadata is updated. If the caller is not registered this call throws NotRegistered.

Args:

‘origin’: (Origin):

The signature of the caller.

‘netuid’ (u16):

The u16 network identifier.

‘version’ (u64):

The bittensor version identifier.

‘ip’ (u64):

The endpoint ip information as a u128 encoded integer.

‘port’ (u16):

The endpoint port information as a u16 encoded integer.

‘ip_type’ (u8):

The endpoint ip version as a u8, 4 or 6.

‘protocol’ (u8):

UDP:1 or TCP:0

‘placeholder1’ (u8):

Placeholder for further extra params.

‘placeholder2’ (u8):

Placeholder for further extra params.

Event:

AxonServed;

On successfully serving the axon info.

Raises:

‘SubNetworkDoesNotExist’:

Attempting to set weights on a non-existent network.

‘NotRegistered’:

Attempting to set weights from a non registered account.

‘InvalidIpType’:

The ip type is not 4 or 6.

‘InvalidIpAddress’:

The numerically encoded ip address does not resolve to a proper ip.

‘ServingRateLimitExceeded’:

Attempting to set prometheus information withing the rate limit min.

Fields

netuid: u16

version: u32

ip: u128

port: u16

ip_type: u8

protocol: u8

placeholder1: u8

placeholder2: u8

serve_axon_tls

Same as serve_axon but takes a certificate as an extra optional argument. Serves or updates axon /prometheus information for the neuron associated with the caller. If the caller is already registered the metadata is updated. If the caller is not registered this call throws NotRegistered.

Args:

‘origin’: (Origin):

The signature of the caller.

‘netuid’ (u16):

The u16 network identifier.

‘version’ (u64):

The bittensor version identifier.

‘ip’ (u64):

The endpoint ip information as a u128 encoded integer.

‘port’ (u16):

The endpoint port information as a u16 encoded integer.

‘ip_type’ (u8):

The endpoint ip version as a u8, 4 or 6.

‘protocol’ (u8):

UDP:1 or TCP:0

‘placeholder1’ (u8):

Placeholder for further extra params.

‘placeholder2’ (u8):

Placeholder for further extra params.

‘certificate’ (Vec):

TLS certificate for inter neuron communitation.

Event:

AxonServed;

On successfully serving the axon info.

Raises:

‘SubNetworkDoesNotExist’:

Attempting to set weights on a non-existent network.

‘NotRegistered’:

Attempting to set weights from a non registered account.

‘InvalidIpType’:

The ip type is not 4 or 6.

‘InvalidIpAddress’:

The numerically encoded ip address does not resolve to a proper ip.

‘ServingRateLimitExceeded’:

Attempting to set prometheus information withing the rate limit min.

Fields

netuid: u16

version: u32

ip: u128

port: u16

ip_type: u8

protocol: u8

placeholder1: u8

placeholder2: u8

certificate: Vec<u8>

serve_prometheus

–– Set prometheus information for the neuron.

Args:

‘origin’: (Origin):

The signature of the calling hotkey.

‘netuid’ (u16):

The u16 network identifier.

‘version’ (u16):

The bittensor version identifier.

‘ip’ (u128):

The prometheus ip information as a u128 encoded integer.

‘port’ (u16):

The prometheus port information as a u16 encoded integer.

‘ip_type’ (u8):

The ip type v4 or v6.

Fields

netuid: u16

version: u32

ip: u128

port: u16

ip_type: u8

register

–– Registers a new neuron to the subnetwork.

Args:

‘origin’: (Origin):

The signature of the calling hotkey.

‘netuid’ (u16):

The u16 network identifier.

‘block_number’ ( u64 ):

Block hash used to prove work done.

‘nonce’ ( u64 ):

Positive integer nonce used in POW.

‘work’ ( Vec ):

Vector encoded bytes representing work done.

‘hotkey’ ( T::AccountId ):

Hotkey to be registered to the network.

‘coldkey’ ( T::AccountId ):

Associated coldkey account.

Event:

NeuronRegistered;

On successfully registering a uid to a neuron slot on a subnetwork.

Raises:

‘SubNetworkDoesNotExist’:

Attempting to register to a non existent network.

‘TooManyRegistrationsThisBlock’:

This registration exceeds the total allowed on this network this block.

‘HotKeyAlreadyRegisteredInSubNet’:

The hotkey is already registered on this network.

‘InvalidWorkBlock’:

The work has been performed on a stale, future, or non existent block.

‘InvalidDifficulty’:

The work does not match the difficulty.

‘InvalidSeal’:

The seal is incorrect.

Fields

netuid: u16

block_number: u64

nonce: u64

work: Vec<u8>

hotkey: T::AccountId

coldkey: T::AccountId

root_register

Register the hotkey to root network

Fields

hotkey: T::AccountId

adjust_senate

Attempt to adjust the senate membership to include a hotkey

Fields

hotkey: T::AccountId

burned_register

User register a new subnetwork via burning token

Fields

netuid: u16

hotkey: T::AccountId

swap_hotkey

The extrinsic for user to change its hotkey

Fields

hotkey: T::AccountId

new_hotkey: T::AccountId

swap_coldkey

The extrinsic for user to change the coldkey associated with their account.

Arguments

origin - The origin of the call, must be signed by the old coldkey.

old_coldkey - The current coldkey associated with the account.

new_coldkey - The new coldkey to be associated with the account.

Returns

Returns a DispatchResultWithPostInfo indicating success or failure of the operation.

Weight

Weight is calculated based on the number of database reads and writes.

Fields

old_coldkey: T::AccountId

new_coldkey: T::AccountId

swap_cost: u64

set_childkey_take

Sets the childkey take for a given hotkey.

This function allows a coldkey to set the childkey take for a given hotkey. The childkey take determines the proportion of stake that the hotkey keeps for itself when distributing stake to its children.

Arguments:

origin (::RuntimeOrigin):

The signature of the calling coldkey. Setting childkey take can only be done by the coldkey.

hotkey (T::AccountId):

The hotkey for which the childkey take will be set.

take (u16):

The new childkey take value. This is a percentage represented as a value between 0 and 10000, where 10000 represents 100%.

Events:

ChildkeyTakeSet:

On successfully setting the childkey take for a hotkey.

Errors:

NonAssociatedColdKey:

The coldkey does not own the hotkey.

InvalidChildkeyTake:

The provided take value is invalid (greater than the maximum allowed take).

TxChildkeyTakeRateLimitExceeded:

The rate limit for changing childkey take has been exceeded.

Fields

hotkey: T::AccountId

netuid: u16

take: u16

sudo_set_tx_childkey_take_rate_limit

Sets the transaction rate limit for changing childkey take.

This function can only be called by the root origin.

Arguments:

origin - The origin of the call, must be root.

tx_rate_limit - The new rate limit in blocks.

Errors:

BadOrigin - If the origin is not root.

Fields

tx_rate_limit: u64

sudo_set_min_childkey_take

Sets the minimum allowed childkey take.

This function can only be called by the root origin.

Arguments:

origin - The origin of the call, must be root.

take - The new minimum childkey take value.

Errors:

BadOrigin - If the origin is not root.

Fields

take: u16

sudo_set_max_childkey_take

Sets the maximum allowed childkey take.

This function can only be called by the root origin.

Arguments:

origin - The origin of the call, must be root.

take - The new maximum childkey take value.

Errors:

BadOrigin - If the origin is not root.

Fields

take: u16

sudo

Authenticates a council proposal and dispatches a function call with Root origin.

The dispatch origin for this call must be a council majority.

Complexity

O(1).

Fields

call: Box<T::SudoRuntimeCall>

sudo_unchecked_weight

Authenticates a council proposal and dispatches a function call with Root origin. This function does not check the weight of the call, and instead allows the user to specify the weight of the call.

The dispatch origin for this call must be a council majority.

Complexity

O(1).

Fields

call: Box<T::SudoRuntimeCall>

weight: Weight

vote

User vote on a proposal

Fields

hotkey: T::AccountId

proposal: T::Hash

index: u32

approve: bool

register_network

User register a new subnetwork

Fields

hotkey: T::AccountId

faucet

Facility extrinsic for user to get taken from faucet It is only available when pow-faucet feature enabled Just deployed in testnet and devnet for testing purpose

Fields

block_number: u64

nonce: u64

work: Vec<u8>

dissolve_network

Remove a user’s subnetwork The caller must be the owner of the network

Fields

coldkey: T::AccountId

netuid: u16

set_children

Set a single child for a given hotkey on a specified network.

This function allows a coldkey to set a single child for a given hotkey on a specified network. The proportion of the hotkey’s stake to be allocated to the child is also specified.

Arguments:

origin (::RuntimeOrigin):

The signature of the calling coldkey. Setting a hotkey child can only be done by the coldkey.

hotkey (T::AccountId):

The hotkey which will be assigned the child.

child (T::AccountId):

The child which will be assigned to the hotkey.

netuid (u16):

The u16 network identifier where the childkey will exist.

proportion (u64):

Proportion of the hotkey’s stake to be given to the child, the value must be u64 normalized.

Events:

ChildAddedSingular:

On successfully registering a child to a hotkey.

Errors:

SubNetworkDoesNotExist:

Attempting to register to a non-existent network.

RegistrationNotPermittedOnRootSubnet:

Attempting to register a child on the root network.

NonAssociatedColdKey:

The coldkey does not own the hotkey or the child is the same as the hotkey.

HotKeyAccountNotExists:

The hotkey account does not exist.

Detailed Explanation of Checks:

Signature Verification: Ensures that the caller has signed the transaction, verifying the coldkey.

Root Network Check: Ensures that the delegation is not on the root network, as child hotkeys are not valid on the root.

Network Existence Check: Ensures that the specified network exists.

Ownership Verification: Ensures that the coldkey owns the hotkey.

Hotkey Account Existence Check: Ensures that the hotkey account already exists.

Child-Hotkey Distinction: Ensures that the child is not the same as the hotkey.

Old Children Cleanup: Removes the hotkey from the parent list of its old children.

New Children Assignment: Assigns the new child to the hotkey and updates the parent list for the new child.

Fields

hotkey: T::AccountId

netuid: u16

children: Vec<(u64, T::AccountId)>

schedule_swap_coldkey

Schedules a coldkey swap operation to be executed at a future block.

This function allows a user to schedule the swapping of their coldkey to a new one at a specified future block. The swap is not executed immediately but is scheduled to occur at the specified block number.

Arguments

origin - The origin of the call, which should be signed by the current coldkey owner.

new_coldkey - The account ID of the new coldkey that will replace the current one.

when - The block number at which the coldkey swap should be executed.

Returns

Returns a DispatchResultWithPostInfo indicating whether the scheduling was successful.

Errors

This function may return an error if:

The origin is not signed.

The scheduling fails due to conflicts or system constraints.

Notes

The actual swap is not performed by this function. It merely schedules the swap operation.

The weight of this call is set to a fixed value and may need adjustment based on benchmarking.

TODO

Implement proper weight calculation based on the complexity of the operation.

Consider adding checks to prevent scheduling too far into the future. TODO: Benchmark this call

Fields

new_coldkey: T::AccountId

schedule_dissolve_network

Schedule the dissolution of a network at a specified block number.

Arguments

origin - The origin of the call, must be signed by the sender.

netuid - The u16 network identifier to be dissolved.

Returns

Returns a DispatchResultWithPostInfo indicating success or failure of the operation.

Weight

Weight is calculated based on the number of database reads and writes.

Fields

netuid: u16

set_identity

–– Set prometheus information for the neuron.

Args:

‘origin’: (Origin):

The signature of the calling hotkey.

‘netuid’ (u16):

The u16 network identifier.

‘version’ (u16):

The bittensor version identifier.

‘ip’ (u128):

The prometheus ip information as a u128 encoded integer.

‘port’ (u16):

The prometheus port information as a u16 encoded integer.

‘ip_type’ (u8):

The ip type v4 or v6.

Fields

name: Vec<u8>

url: Vec<u8>

github_repo: Vec<u8>

image: Vec<u8>

discord: Vec<u8>

description: Vec<u8>

additional: Vec<u8>

set_subnet_identity

–– Set the identity information for a subnet.

Args:

origin - (::Origin):

The signature of the calling coldkey, which must be the owner of the subnet.

netuid (u16):

The unique network identifier of the subnet.

subnet_name (Vec):

The name of the subnet.

github_repo (Vec):

The GitHub repository associated with the subnet identity.

subnet_contact (Vec):

The contact information for the subnet.

Fields

netuid: u16

subnet_name: Vec<u8>

github_repo: Vec<u8>

subnet_contact: Vec<u8>

subnet_url: Vec<u8>

discord: Vec<u8>

description: Vec<u8>

additional: Vec<u8>

register_network_with_identity

User register a new subnetwork

Fields

hotkey: T::AccountId

identity: Option<SubnetIdentityOfV2>

unstake_all

–– The implementation for the extrinsic unstake_all: Removes all stake from a hotkey account across all subnets and adds it onto a coldkey.

Args:

origin - (::Origin):

The signature of the caller’s coldkey.

hotkey (T::AccountId):

The associated hotkey account.

Event:

StakeRemoved;

On the successfully removing stake from the hotkey account.

Raises:

NotRegistered:

Thrown if the account we are attempting to unstake from is non existent.

NonAssociatedColdKey:

Thrown if the coldkey does not own the hotkey we are unstaking from.

NotEnoughStakeToWithdraw:

Thrown if there is not enough stake on the hotkey to withdraw this amount.

TxRateLimitExceeded:

Thrown if key has hit transaction rate limit

Fields

hotkey: T::AccountId

unstake_all_alpha

–– The implementation for the extrinsic unstake_all: Removes all stake from a hotkey account across all subnets and adds it onto a coldkey.

Args:

origin - (::Origin):

The signature of the caller’s coldkey.

hotkey (T::AccountId):

The associated hotkey account.

Event:

StakeRemoved;

On the successfully removing stake from the hotkey account.

Raises:

NotRegistered:

Thrown if the account we are attempting to unstake from is non existent.

NonAssociatedColdKey:

Thrown if the coldkey does not own the hotkey we are unstaking from.

NotEnoughStakeToWithdraw:

Thrown if there is not enough stake on the hotkey to withdraw this amount.

TxRateLimitExceeded:

Thrown if key has hit transaction rate limit

Fields

hotkey: T::AccountId

move_stake

–– The implementation for the extrinsic move_stake: Moves specified amount of stake from a hotkey to another across subnets.

Args:

origin - (::Origin):

The signature of the caller’s coldkey.

origin_hotkey (T::AccountId):

The hotkey account to move stake from.

destination_hotkey (T::AccountId):

The hotkey account to move stake to.

origin_netuid (T::AccountId):

The subnet ID to move stake from.

destination_netuid (T::AccountId):

The subnet ID to move stake to.

alpha_amount (T::AccountId):

The alpha stake amount to move.

Fields

origin_hotkey: T::AccountId

destination_hotkey: T::AccountId

origin_netuid: u16

destination_netuid: u16

alpha_amount: u64

transfer_stake

Transfers a specified amount of stake from one coldkey to another, optionally across subnets, while keeping the same hotkey.

Arguments

origin - The origin of the transaction, which must be signed by the origin_coldkey.

destination_coldkey - The coldkey to which the stake is transferred.

hotkey - The hotkey associated with the stake.

origin_netuid - The network/subnet ID to move stake from.

destination_netuid - The network/subnet ID to move stake to (for cross-subnet transfer).

alpha_amount - The amount of stake to transfer.

Errors

Returns an error if:

The origin is not signed by the correct coldkey.

Either subnet does not exist.

The hotkey does not exist.

There is insufficient stake on (origin_coldkey, hotkey, origin_netuid).

The transfer amount is below the minimum stake requirement.

Events

May emit a StakeTransferred event on success.

Fields

destination_coldkey: T::AccountId

hotkey: T::AccountId

origin_netuid: u16

destination_netuid: u16

alpha_amount: u64

swap_stake

Swaps a specified amount of stake from one subnet to another, while keeping the same coldkey and hotkey.

Arguments

origin - The origin of the transaction, which must be signed by the coldkey that owns the hotkey.

hotkey - The hotkey whose stake is being swapped.

origin_netuid - The network/subnet ID from which stake is removed.

destination_netuid - The network/subnet ID to which stake is added.

alpha_amount - The amount of stake to swap.

Errors

Returns an error if:

The transaction is not signed by the correct coldkey (i.e., coldkey_owns_hotkey fails).

Either origin_netuid or destination_netuid does not exist.

The hotkey does not exist.

There is insufficient stake on (coldkey, hotkey, origin_netuid).

The swap amount is below the minimum stake requirement.

Events

May emit a StakeSwapped event on success.

Fields

hotkey: T::AccountId

origin_netuid: u16

destination_netuid: u16

alpha_amount: u64

add_stake_limit

— Adds stake to a hotkey on a subnet with a price limit. This extrinsic allows to specify the limit price for alpha token at which or better (lower) the staking should execute.

In case if slippage occurs and the price shall move beyond the limit price, the staking order may execute only partially or not execute at all.

Args:

‘origin’: (Origin):

The signature of the caller’s coldkey.

‘hotkey’ (T::AccountId):

The associated hotkey account.

‘netuid’ (u16):

Subnetwork UID

‘amount_staked’ (u64):

The amount of stake to be added to the hotkey staking account.

‘limit_price’ (u64):

The limit price expressed in units of RAO per one Alpha.

‘allow_partial’ (bool):

Allows partial execution of the amount. If set to false, this becomes fill or kill type or order.

Event:

StakeAdded;

On the successfully adding stake to a global account.

Raises:

‘NotEnoughBalanceToStake’:

Not enough balance on the coldkey to add onto the global account.

‘NonAssociatedColdKey’:

The calling coldkey is not associated with this hotkey.

‘BalanceWithdrawalError’:

Errors stemming from transaction pallet.

Fields

hotkey: T::AccountId

netuid: u16

amount_staked: u64

limit_price: u64

allow_partial: bool

remove_stake_limit

— Removes stake from a hotkey on a subnet with a price limit. This extrinsic allows to specify the limit price for alpha token at which or better (higher) the staking should execute.

In case if slippage occurs and the price shall move beyond the limit price, the staking order may execute only partially or not execute at all.

Args:

‘origin’: (Origin):

The signature of the caller’s coldkey.

‘hotkey’ (T::AccountId):

The associated hotkey account.

‘netuid’ (u16):

Subnetwork UID

‘amount_unstaked’ (u64):

The amount of stake to be added to the hotkey staking account.

‘limit_price’ (u64):

The limit price expressed in units of RAO per one Alpha.

‘allow_partial’ (bool):

Allows partial execution of the amount. If set to false, this becomes fill or kill type or order.

Event:

StakeRemoved;

On the successfully removing stake from the hotkey account.

Raises:

‘NotRegistered’:

Thrown if the account we are attempting to unstake from is non existent.

‘NonAssociatedColdKey’:

Thrown if the coldkey does not own the hotkey we are unstaking from.

‘NotEnoughStakeToWithdraw’:

Thrown if there is not enough stake on the hotkey to withdwraw this amount.

Fields

hotkey: T::AccountId

netuid: u16

amount_unstaked: u64

limit_price: u64

allow_partial: bool

swap_stake_limit

Swaps a specified amount of stake from one subnet to another, while keeping the same coldkey and hotkey.

Arguments

origin - The origin of the transaction, which must be signed by the coldkey that owns the hotkey.

hotkey - The hotkey whose stake is being swapped.

origin_netuid - The network/subnet ID from which stake is removed.

destination_netuid - The network/subnet ID to which stake is added.

alpha_amount - The amount of stake to swap.

limit_price - The limit price expressed in units of RAO per one Alpha.

allow_partial - Allows partial execution of the amount. If set to false, this becomes fill or kill type or order.

Errors

Returns an error if:

The transaction is not signed by the correct coldkey (i.e., coldkey_owns_hotkey fails).

Either origin_netuid or destination_netuid does not exist.

The hotkey does not exist.

There is insufficient stake on (coldkey, hotkey, origin_netuid).

The swap amount is below the minimum stake requirement.

Events

May emit a StakeSwapped event on success.

Fields

hotkey: T::AccountId

origin_netuid: u16

destination_netuid: u16

alpha_amount: u64

limit_price: u64

allow_partial: bool

try_associate_hotkey

Attempts to associate a hotkey with a coldkey.

Arguments

origin - The origin of the transaction, which must be signed by the coldkey that owns the hotkey.

hotkey - The hotkey to associate with the coldkey.

Note

Will charge based on the weight even if the hotkey is already associated with a coldkey.

Fields

hotkey: T::AccountId

start_call

Initiates a call on a subnet.

Arguments

origin - The origin of the call, which must be signed by the subnet owner.

netuid - The unique identifier of the subnet on which the call is being initiated.

Events

Emits a FirstEmissionBlockNumberSet event on success.

Fields

netuid: u16

associate_evm_key

Attempts to associate a hotkey with an EVM key.

The signature will be checked to see if the recovered public key matches the evm_key provided.

The EVM key is expected to sign the message according to this formula to produce the signature:

keccak_256(hotkey ++ keccak_256(block_number))

Arguments

origin - The origin of the transaction, which must be signed by the coldkey that owns the hotkey.

netuid - The netuid that the hotkey belongs to.

hotkey - The hotkey associated with the origin.

evm_key - The EVM key to associate with the hotkey.

block_number - The block number used in the signature.

signature - A signed message by the evm_key containing the hotkey and the hashed block_number.

Errors

Returns an error if:

The transaction is not signed.

The hotkey is not owned by the origin coldkey.

The hotkey does not belong to the subnet identified by the netuid.

The EVM key cannot be recovered from the signature.

The EVM key recovered from the signature does not match the given EVM key.

Events

May emit a EvmKeyAssociated event on success

Fields

netuid: u16

hotkey: T::AccountId

evm_key: H160

block_number: u64

signature: Signature

recycle_alpha

Recycles alpha from a cold/hot key pair, reducing AlphaOut on a subnet

Arguments

origin - The origin of the call (must be signed by the coldkey)

hotkey - The hotkey account

amount - The amount of alpha to recycle

netuid - The subnet ID

Events

Emits a TokensRecycled event on success.

Fields

hotkey: T::AccountId

amount: u64

netuid: u16

burn_alpha

Burns alpha from a cold/hot key pair without reducing AlphaOut

Arguments

origin - The origin of the call (must be signed by the coldkey)

hotkey - The hotkey account

amount - The amount of alpha to burn

netuid - The subnet ID

Events

Emits a TokensBurned event on success.

Fields

hotkey: T::AccountId

amount: u64

netuid: u16

Implementations

impl<T: Config> Call<T>

pub fn new_call_variant_set_weights( netuid: u16, dests: Vec<u16>, weights: Vec<u16>, version_key: u64, ) -> Self

Create a call with the variant set_weights.

pub fn new_call_variant_batch_set_weights( netuids: Vec<Compact<u16>>, weights: Vec<Vec<(Compact<u16>, Compact<u16>)>>, version_keys: Vec<Compact<u64>>, ) -> Self

Create a call with the variant batch_set_weights.

pub fn new_call_variant_commit_weights(netuid: u16, commit_hash: H256) -> Self

Create a call with the variant commit_weights.

pub fn new_call_variant_batch_commit_weights( netuids: Vec<Compact<u16>>, commit_hashes: Vec<H256>, ) -> Self

Create a call with the variant batch_commit_weights.

pub fn new_call_variant_reveal_weights( netuid: u16, uids: Vec<u16>, values: Vec<u16>, salt: Vec<u16>, version_key: u64, ) -> Self

Create a call with the variant reveal_weights.

pub fn new_call_variant_commit_crv3_weights( netuid: u16, commit: BoundedVec<u8, ConstU32<MAX_CRV3_COMMIT_SIZE_BYTES>>, reveal_round: u64, ) -> Self

Create a call with the variant commit_crv3_weights.

pub fn new_call_variant_batch_reveal_weights( netuid: u16, uids_list: Vec<Vec<u16>>, values_list: Vec<Vec<u16>>, salts_list: Vec<Vec<u16>>, version_keys: Vec<u64>, ) -> Self

Create a call with the variant batch_reveal_weights.

pub fn new_call_variant_set_tao_weights( netuid: u16, hotkey: T::AccountId, dests: Vec<u16>, weights: Vec<u16>, version_key: u64, ) -> Self

Create a call with the variant set_tao_weights.

pub fn new_call_variant_become_delegate(hotkey: T::AccountId) -> Self

Create a call with the variant become_delegate.

pub fn new_call_variant_decrease_take(hotkey: T::AccountId, take: u16) -> Self

Create a call with the variant decrease_take.

pub fn new_call_variant_increase_take(hotkey: T::AccountId, take: u16) -> Self

Create a call with the variant increase_take.

pub fn new_call_variant_add_stake( hotkey: T::AccountId, netuid: u16, amount_staked: u64, ) -> Self

Create a call with the variant add_stake.

pub fn new_call_variant_remove_stake( hotkey: T::AccountId, netuid: u16, amount_unstaked: u64, ) -> Self

Create a call with the variant remove_stake.

pub fn new_call_variant_serve_axon( netuid: u16, version: u32, ip: u128, port: u16, ip_type: u8, protocol: u8, placeholder1: u8, placeholder2: u8, ) -> Self

Create a call with the variant serve_axon.

pub fn new_call_variant_serve_axon_tls( netuid: u16, version: u32, ip: u128, port: u16, ip_type: u8, protocol: u8, placeholder1: u8, placeholder2: u8, certificate: Vec<u8>, ) -> Self

Create a call with the variant serve_axon_tls.

pub fn new_call_variant_serve_prometheus( netuid: u16, version: u32, ip: u128, port: u16, ip_type: u8, ) -> Self

Create a call with the variant serve_prometheus.

pub fn new_call_variant_register( netuid: u16, block_number: u64, nonce: u64, work: Vec<u8>, hotkey: T::AccountId, coldkey: T::AccountId, ) -> Self

Create a call with the variant register.

pub fn new_call_variant_root_register(hotkey: T::AccountId) -> Self

Create a call with the variant root_register.

pub fn new_call_variant_adjust_senate(hotkey: T::AccountId) -> Self

Create a call with the variant adjust_senate.

pub fn new_call_variant_burned_register( netuid: u16, hotkey: T::AccountId, ) -> Self

Create a call with the variant burned_register.

pub fn new_call_variant_swap_hotkey( hotkey: T::AccountId, new_hotkey: T::AccountId, ) -> Self

Create a call with the variant swap_hotkey.

pub fn new_call_variant_swap_coldkey( old_coldkey: T::AccountId, new_coldkey: T::AccountId, swap_cost: u64, ) -> Self

Create a call with the variant swap_coldkey.

pub fn new_call_variant_set_childkey_take( hotkey: T::AccountId, netuid: u16, take: u16, ) -> Self

Create a call with the variant set_childkey_take.

pub fn new_call_variant_sudo_set_tx_childkey_take_rate_limit( tx_rate_limit: u64, ) -> Self

Create a call with the variant sudo_set_tx_childkey_take_rate_limit.

pub fn new_call_variant_sudo_set_min_childkey_take(take: u16) -> Self

Create a call with the variant sudo_set_min_childkey_take.

pub fn new_call_variant_sudo_set_max_childkey_take(take: u16) -> Self

Create a call with the variant sudo_set_max_childkey_take.

pub fn new_call_variant_sudo(call: Box<T::SudoRuntimeCall>) -> Self

Create a call with the variant sudo.

pub fn new_call_variant_sudo_unchecked_weight( call: Box<T::SudoRuntimeCall>, weight: Weight, ) -> Self

Create a call with the variant sudo_unchecked_weight.

pub fn new_call_variant_vote( hotkey: T::AccountId, proposal: T::Hash, index: u32, approve: bool, ) -> Self

Create a call with the variant vote.

pub fn new_call_variant_register_network(hotkey: T::AccountId) -> Self

Create a call with the variant register_network.

pub fn new_call_variant_faucet( block_number: u64, nonce: u64, work: Vec<u8>, ) -> Self

Create a call with the variant faucet.

pub fn new_call_variant_dissolve_network( coldkey: T::AccountId, netuid: u16, ) -> Self

Create a call with the variant dissolve_network.

pub fn new_call_variant_set_children( hotkey: T::AccountId, netuid: u16, children: Vec<(u64, T::AccountId)>, ) -> Self

Create a call with the variant set_children.

pub fn new_call_variant_schedule_swap_coldkey(new_coldkey: T::AccountId) -> Self

Create a call with the variant schedule_swap_coldkey.

pub fn new_call_variant_schedule_dissolve_network(netuid: u16) -> Self

Create a call with the variant schedule_dissolve_network.

pub fn new_call_variant_set_identity( name: Vec<u8>, url: Vec<u8>, github_repo: Vec<u8>, image: Vec<u8>, discord: Vec<u8>, description: Vec<u8>, additional: Vec<u8>, ) -> Self

Create a call with the variant set_identity.

pub fn new_call_variant_set_subnet_identity( netuid: u16, subnet_name: Vec<u8>, github_repo: Vec<u8>, subnet_contact: Vec<u8>, subnet_url: Vec<u8>, discord: Vec<u8>, description: Vec<u8>, additional: Vec<u8>, ) -> Self

Create a call with the variant set_subnet_identity.

pub fn new_call_variant_register_network_with_identity( hotkey: T::AccountId, identity: Option<SubnetIdentityOfV2>, ) -> Self

Create a call with the variant register_network_with_identity.

pub fn new_call_variant_unstake_all(hotkey: T::AccountId) -> Self

Create a call with the variant unstake_all.

pub fn new_call_variant_unstake_all_alpha(hotkey: T::AccountId) -> Self

Create a call with the variant unstake_all_alpha.

pub fn new_call_variant_move_stake( origin_hotkey: T::AccountId, destination_hotkey: T::AccountId, origin_netuid: u16, destination_netuid: u16, alpha_amount: u64, ) -> Self

Create a call with the variant move_stake.

pub fn new_call_variant_transfer_stake( destination_coldkey: T::AccountId, hotkey: T::AccountId, origin_netuid: u16, destination_netuid: u16, alpha_amount: u64, ) -> Self

Create a call with the variant transfer_stake.

pub fn new_call_variant_swap_stake( hotkey: T::AccountId, origin_netuid: u16, destination_netuid: u16, alpha_amount: u64, ) -> Self

Create a call with the variant swap_stake.

pub fn new_call_variant_add_stake_limit( hotkey: T::AccountId, netuid: u16, amount_staked: u64, limit_price: u64, allow_partial: bool, ) -> Self

Create a call with the variant add_stake_limit.

pub fn new_call_variant_remove_stake_limit( hotkey: T::AccountId, netuid: u16, amount_unstaked: u64, limit_price: u64, allow_partial: bool, ) -> Self

Create a call with the variant remove_stake_limit.

pub fn new_call_variant_swap_stake_limit( hotkey: T::AccountId, origin_netuid: u16, destination_netuid: u16, alpha_amount: u64, limit_price: u64, allow_partial: bool, ) -> Self

Create a call with the variant swap_stake_limit.

pub fn new_call_variant_try_associate_hotkey(hotkey: T::AccountId) -> Self

Create a call with the variant try_associate_hotkey.

pub fn new_call_variant_start_call(netuid: u16) -> Self

Create a call with the variant start_call.

pub fn new_call_variant_associate_evm_key( netuid: u16, hotkey: T::AccountId, evm_key: H160, block_number: u64, signature: Signature, ) -> Self

Create a call with the variant associate_evm_key.

pub fn new_call_variant_recycle_alpha( hotkey: T::AccountId, amount: u64, netuid: u16, ) -> Self

Create a call with the variant recycle_alpha.

pub fn new_call_variant_burn_alpha( hotkey: T::AccountId, amount: u64, netuid: u16, ) -> Self

Create a call with the variant burn_alpha.