pallet_subtensor::pallet

Enum Error

pub enum Error<T> {

    SubNetworkDoesNotExist,
    RootNetworkDoesNotExist,
    InvalidIpType,
    InvalidIpAddress,
    InvalidPort,
    HotKeyNotRegisteredInSubNet,
    HotKeyAccountNotExists,
    HotKeyNotRegisteredInNetwork,
    NonAssociatedColdKey,
    NotEnoughStake,
    NotEnoughStakeToWithdraw,
    NotEnoughStakeToSetWeights,
    NotEnoughStakeToSetChildkeys,
    NotEnoughBalanceToStake,
    BalanceWithdrawalError,
    ZeroBalanceAfterWithdrawn,
    NeuronNoValidatorPermit,
    WeightVecNotEqualSize,
    DuplicateUids,
    UidVecContainInvalidOne,
    WeightVecLengthIsLow,
    TooManyRegistrationsThisBlock,
    HotKeyAlreadyRegisteredInSubNet,
    NewHotKeyIsSameWithOld,
    InvalidWorkBlock,
    InvalidDifficulty,
    InvalidSeal,
    MaxWeightExceeded,
    HotKeyAlreadyDelegate,
    SettingWeightsTooFast,
    IncorrectWeightVersionKey,
    ServingRateLimitExceeded,
    UidsLengthExceedUidsInSubNet,
    NetworkTxRateLimitExceeded,
    DelegateTxRateLimitExceeded,
    HotKeySetTxRateLimitExceeded,
    StakingRateLimitExceeded,
    SubNetRegistrationDisabled,
    TooManyRegistrationsThisInterval,
    TransactorAccountShouldBeHotKey,
    NotSenateMember,
    FaucetDisabled,
    NotSubnetOwner,
    RegistrationNotPermittedOnRootSubnet,
    StakeTooLowForRoot,
    AllNetworksInImmunity,
    NotEnoughBalanceToPaySwapHotKey,
    NotRootSubnet,
    CanNotSetRootNetworkWeights,
    NoNeuronIdAvailable,
    DelegateTakeTooLow,
    DelegateTakeTooHigh,
    NoWeightsCommitFound,
    InvalidRevealCommitHashNotMatch,
    CommitRevealEnabled,
    CommitRevealDisabled,
    CouldNotJoinSenate,
    LiquidAlphaDisabled,
    AlphaHighTooLow,
    AlphaLowOutOfRange,
    ColdKeyAlreadyAssociated,
    NotEnoughBalanceToPaySwapColdKey,
    ColdkeyIsInArbitration,
    InvalidChild,
    DuplicateChild,
    ProportionOverflow,
    TooManyChildren,
    TxRateLimitExceeded,
    SwapAlreadyScheduled,
    FailedToSchedule,
    NewColdKeyIsHotkey,
    InvalidChildkeyTake,
    TxChildkeyTakeRateLimitExceeded,
    InvalidIdentity,
    MechanismDoesNotExist,
    CannotUnstakeLock,
    SubnetNotExists,
    TooManyUnrevealedCommits,
    ExpiredWeightCommit,
    RevealTooEarly,
    InputLengthsUnequal,
    CommittingWeightsTooFast,
    AmountTooLow,
    InsufficientLiquidity,
    SlippageTooHigh,
    TransferDisallowed,
    ActivityCutoffTooLow,
    CallDisabled,
    FirstEmissionBlockNumberAlreadySet,
    NeedWaitingMoreBlocksToStarCall,
    NotEnoughAlphaOutToRecycle,
    CannotBurnOrRecycleOnRootSubnet,
    UnableToRecoverPublicKey,
    InvalidRecoveredPublicKey,
    SubtokenDisabled,
    ZeroMaxStakeAmount,
    SameNetuid,
    // some variants omitted

}

The Error enum of this pallet.

Variants

SubNetworkDoesNotExist

The subnet does not exist.

RootNetworkDoesNotExist

The root network does not exist.

InvalidIpType

The user is trying to serve an axon which is not of type 4 (IPv4) or 6 (IPv6).

InvalidIpAddress

An invalid IP address is passed to the serve function.

InvalidPort

An invalid port is passed to the serve function.

HotKeyNotRegisteredInSubNet

The hotkey is not registered in subnet

HotKeyAccountNotExists

The hotkey does not exists

HotKeyNotRegisteredInNetwork

The hotkey is not registered in any subnet.

NonAssociatedColdKey

Request to stake, unstake or subscribe is made by a coldkey that is not associated with the hotkey account.

NotEnoughStake

DEPRECATED: Stake amount to withdraw is zero. The caller does not have enought stake to perform this action.

NotEnoughStakeToWithdraw

The caller is requesting removing more stake than there exists in the staking account. See: “[remove_stake()]”.

NotEnoughStakeToSetWeights

The caller is requesting to set weights but the caller has less than minimum stake required to set weights (less than WeightsMinStake).

NotEnoughStakeToSetChildkeys

The parent hotkey doesn’t have enough own stake to set childkeys.

NotEnoughBalanceToStake

The caller is requesting adding more stake than there exists in the coldkey account. See: “[add_stake()]”

BalanceWithdrawalError

The caller is trying to add stake, but for some reason the requested amount could not be withdrawn from the coldkey account.

ZeroBalanceAfterWithdrawn

Unsuccessfully withdraw, balance could be zero (can not make account exist) after withdrawal.

NeuronNoValidatorPermit

The caller is attempting to set non-self weights without being a permitted validator.

WeightVecNotEqualSize

The caller is attempting to set the weight keys and values but these vectors have different size.

DuplicateUids

The caller is attempting to set weights with duplicate UIDs in the weight matrix.

UidVecContainInvalidOne

The caller is attempting to set weight to at least one UID that does not exist in the metagraph.

WeightVecLengthIsLow

The dispatch is attempting to set weights on chain with fewer elements than are allowed.

TooManyRegistrationsThisBlock

Number of registrations in this block exceeds the allowed number (i.e., exceeds the subnet hyperparameter “max_regs_per_block”).

HotKeyAlreadyRegisteredInSubNet

The caller is requesting registering a neuron which already exists in the active set.

NewHotKeyIsSameWithOld

The new hotkey is the same as old one

InvalidWorkBlock

The supplied PoW hash block is in the future or negative.

InvalidDifficulty

The supplied PoW hash block does not meet the network difficulty.

InvalidSeal

The supplied PoW hash seal does not match the supplied work.

MaxWeightExceeded

The dispatch is attempting to set weights on chain with weight value exceeding the MaxWeightLimit (max_weight_limit subnet hyperparameter).

HotKeyAlreadyDelegate

The hotkey is attempting to become a delegate when the hotkey is already a delegate.

SettingWeightsTooFast

A transactor exceeded the rate limit for setting weights.

IncorrectWeightVersionKey

A validator is attempting to set weights from a validator with incorrect weight version.

ServingRateLimitExceeded

An axon or prometheus serving exceeded the rate limit for a registered neuron.

UidsLengthExceedUidsInSubNet

The caller is attempting to set weights with more UIDs than allowed.

NetworkTxRateLimitExceeded

A transactor exceeded the rate limit for add network transaction.

DelegateTxRateLimitExceeded

A transactor exceeded the rate limit for delegate transaction.

HotKeySetTxRateLimitExceeded

A transactor exceeded the rate limit for setting or swapping hotkey.

StakingRateLimitExceeded

A transactor exceeded the rate limit for staking.

SubNetRegistrationDisabled

Registration is disabled.

TooManyRegistrationsThisInterval

The number of registration attempts exceeded the allowed number in the interval.

TransactorAccountShouldBeHotKey

The hotkey is required to be the origin.

NotSenateMember

A hotkey is attempting to do something only senate members can do.

FaucetDisabled

Faucet is disabled.

NotSubnetOwner

Not a subnet owner.

RegistrationNotPermittedOnRootSubnet

Operation is not permitted on the root subnet.

StakeTooLowForRoot

A hotkey with too little stake is attempting to join the root subnet.

AllNetworksInImmunity

All subnets are in the immunity period.

NotEnoughBalanceToPaySwapHotKey

Not enough balance to pay swapping hotkey.

NotRootSubnet

Netuid does not match for setting root network weights.

CanNotSetRootNetworkWeights

Can not set weights for the root network.

NoNeuronIdAvailable

No neuron ID is available.

DelegateTakeTooLow

Delegate take is too low.

DelegateTakeTooHigh

Delegate take is too high.

NoWeightsCommitFound

No commit found for the provided hotkey+netuid combination when attempting to reveal the weights.

InvalidRevealCommitHashNotMatch

Committed hash does not equal the hashed reveal data.

CommitRevealEnabled

Attempting to call set_weights when commit/reveal is enabled

CommitRevealDisabled

Attemtping to commit/reveal weights when disabled.

CouldNotJoinSenate

Not able to join the senate.

LiquidAlphaDisabled

Attempting to set alpha high/low while disabled

AlphaHighTooLow

Alpha high is too low: alpha_high > 0.8

AlphaLowOutOfRange

Alpha low is out of range: alpha_low > 0 && alpha_low < 0.8

ColdKeyAlreadyAssociated

The coldkey has already been swapped

NotEnoughBalanceToPaySwapColdKey

The coldkey balance is not enough to pay for the swap

ColdkeyIsInArbitration

The coldkey is in arbitration

InvalidChild

Attempting to set an invalid child for a hotkey on a network.

DuplicateChild

Duplicate child when setting children.

ProportionOverflow

Proportion overflow when setting children.

TooManyChildren

Too many children MAX 5.

TxRateLimitExceeded

Default transaction rate limit exceeded.

SwapAlreadyScheduled

Swap already scheduled.

FailedToSchedule

failed to swap coldkey

NewColdKeyIsHotkey

New coldkey is hotkey

InvalidChildkeyTake

Childkey take is invalid.

TxChildkeyTakeRateLimitExceeded

Childkey take rate limit exceeded.

InvalidIdentity

Invalid identity.

MechanismDoesNotExist

Trying to register a subnet into a mechanism that does not exist.

CannotUnstakeLock

Trying to unstake your lock amount.

SubnetNotExists

Trying to perform action on non-existent subnet.

TooManyUnrevealedCommits

Maximum commit limit reached

ExpiredWeightCommit

Attempted to reveal weights that are expired.

RevealTooEarly

Attempted to reveal weights too early.

InputLengthsUnequal

Attempted to batch reveal weights with mismatched vector input lenghts.

CommittingWeightsTooFast

A transactor exceeded the rate limit for setting weights.

AmountTooLow

Stake amount is too low.

InsufficientLiquidity

Not enough liquidity.

SlippageTooHigh

Slippage is too high for the transaction.

TransferDisallowed

Subnet disallows transfer.

ActivityCutoffTooLow

Activity cutoff is being set too low.

CallDisabled

Call is disabled

FirstEmissionBlockNumberAlreadySet

FirstEmissionBlockNumber is already set.

NeedWaitingMoreBlocksToStarCall

need wait for more blocks to accept the start call extrinsic.

NotEnoughAlphaOutToRecycle

Not enough AlphaOut on the subnet to recycle

CannotBurnOrRecycleOnRootSubnet

Cannot burn or recycle TAO from root subnet

UnableToRecoverPublicKey

Public key cannot be recovered.

InvalidRecoveredPublicKey

Recovered public key is invalid.

SubtokenDisabled

SubToken disabled now

ZeroMaxStakeAmount

Zero max stake amount

SameNetuid

Invalid netuid duplication