pallet_subtensor::pallet

Enum Event

pub enum Event<T: Config> {

    NetworkAdded(u16, u16),
    NetworkRemoved(u16),
    StakeAdded(T::AccountId, T::AccountId, u64, u64, u16, u64),
    StakeRemoved(T::AccountId, T::AccountId, u64, u64, u16, u64),
    AggregatedStakeAdded(T::AccountId, T::AccountId, u16, u64),
    FailedToAddAggregatedStake(T::AccountId, T::AccountId, u16, u64),
    AggregatedLimitedStakeAdded(T::AccountId, T::AccountId, u16, u64, u64, bool),
    FailedToAddAggregatedLimitedStake(T::AccountId, T::AccountId, u16, u64, u64, bool),
    AggregatedStakeRemoved(T::AccountId, T::AccountId, u16, u64),
    FailedToRemoveAggregatedStake(T::AccountId, T::AccountId, u16, u64),
    AggregatedLimitedStakeRemoved(T::AccountId, T::AccountId, u16, u64, u64, bool),
    FailedToRemoveAggregatedLimitedStake(T::AccountId, T::AccountId, u16, u64, u64, bool),
    AggregatedUnstakeAllSucceeded(T::AccountId, T::AccountId),
    AggregatedUnstakeAllFailed(T::AccountId, T::AccountId),
    AggregatedUnstakeAllAlphaSucceeded(T::AccountId, T::AccountId),
    AggregatedUnstakeAllAlphaFailed(T::AccountId, T::AccountId),
    StakeMoved(T::AccountId, T::AccountId, u16, T::AccountId, u16, u64),
    WeightsSet(u16, u16),
    NeuronRegistered(u16, u16, T::AccountId),
    BulkNeuronsRegistered(u16, u16),
    BulkBalancesSet(u16, u16),
    MaxAllowedUidsSet(u16, u16),
    MaxWeightLimitSet(u16, u16),
    DifficultySet(u16, u64),
    AdjustmentIntervalSet(u16, u16),
    RegistrationPerIntervalSet(u16, u16),
    MaxRegistrationsPerBlockSet(u16, u16),
    ActivityCutoffSet(u16, u16),
    RhoSet(u16, u16),
    AlphaSigmoidSteepnessSet(u16, u16),
    KappaSet(u16, u16),
    MinAllowedWeightSet(u16, u16),
    ValidatorPruneLenSet(u16, u64),
    ScalingLawPowerSet(u16, u16),
    WeightsSetRateLimitSet(u16, u64),
    ImmunityPeriodSet(u16, u16),
    BondsMovingAverageSet(u16, u64),
    BondsPenaltySet(u16, u16),
    BondsResetOnSet(u16, bool),
    MaxAllowedValidatorsSet(u16, u16),
    AxonServed(u16, T::AccountId),
    PrometheusServed(u16, T::AccountId),
    DelegateAdded(T::AccountId, T::AccountId, u16),
    DefaultTakeSet(u16),
    WeightsVersionKeySet(u16, u64),
    MinDifficultySet(u16, u64),
    MaxDifficultySet(u16, u64),
    ServingRateLimitSet(u16, u64),
    BurnSet(u16, u64),
    MaxBurnSet(u16, u64),
    MinBurnSet(u16, u64),
    TxRateLimitSet(u64),
    TxDelegateTakeRateLimitSet(u64),
    TxChildKeyTakeRateLimitSet(u64),
    MinChildKeyTakeSet(u16),
    MaxChildKeyTakeSet(u16),
    ChildKeyTakeSet(T::AccountId, u16),
    Sudid(DispatchResult),
    RegistrationAllowed(u16, bool),
    PowRegistrationAllowed(u16, bool),
    TempoSet(u16, u16),
    RAORecycledForRegistrationSet(u16, u64),
    StakeThresholdSet(u64),
    SenateRequiredStakePercentSet(u64),
    AdjustmentAlphaSet(u16, u64),
    Faucet(T::AccountId, u64),
    SubnetOwnerCutSet(u16),
    NetworkRateLimitSet(u64),
    NetworkImmunityPeriodSet(u64),
    NetworkMinLockCostSet(u64),
    NetworkLockCostReductionIntervalSet(u64),
    TakeDecreased(T::AccountId, T::AccountId, u16),
    TakeIncreased(T::AccountId, T::AccountId, u16),
    HotkeySwapped {
        coldkey: T::AccountId,
        old_hotkey: T::AccountId,
        new_hotkey: T::AccountId,
    },
    MaxDelegateTakeSet(u16),
    MinDelegateTakeSet(u16),
    SenateAdjusted {
        old_member: Option<T::AccountId>,
        new_member: T::AccountId,
    },
    ColdkeySwapped {
        old_coldkey: T::AccountId,
        new_coldkey: T::AccountId,
        swap_cost: u64,
    },
    AllBalanceUnstakedAndTransferredToNewColdkey {
        current_coldkey: T::AccountId,
        new_coldkey: T::AccountId,
        total_balance: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance,
    },
    ColdkeySwapScheduled {
        old_coldkey: T::AccountId,
        new_coldkey: T::AccountId,
        execution_block: BlockNumberFor<T>,
        swap_cost: u64,
    },
    ArbitrationPeriodExtended {
        coldkey: T::AccountId,
    },
    SetChildrenScheduled(T::AccountId, u16, u64, Vec<(u64, T::AccountId)>),
    SetChildren(T::AccountId, u16, Vec<(u64, T::AccountId)>),
    ChainIdentitySet(T::AccountId),
    SubnetIdentitySet(u16),
    SubnetIdentityRemoved(u16),
    DissolveNetworkScheduled {
        account: T::AccountId,
        netuid: u16,
        execution_block: BlockNumberFor<T>,
    },
    ColdkeySwapScheduleDurationSet(BlockNumberFor<T>),
    DissolveNetworkScheduleDurationSet(BlockNumberFor<T>),
    CRV3WeightsCommitted(T::AccountId, u16, H256),
    WeightsCommitted(T::AccountId, u16, H256),
    WeightsRevealed(T::AccountId, u16, H256),
    WeightsBatchRevealed(T::AccountId, u16, Vec<H256>),
    BatchWeightsCompleted(Vec<Compact<u16>>, T::AccountId),
    BatchCompletedWithErrors(),
    BatchWeightItemFailed(DispatchError),
    StakeTransferred(T::AccountId, T::AccountId, T::AccountId, u16, u16, u64),
    StakeSwapped(T::AccountId, T::AccountId, u16, u16, u64),
    TransferToggle(u16, bool),
    SubnetOwnerHotkeySet(u16, T::AccountId),
    FirstEmissionBlockNumberSet(u16, u64),
    AlphaRecycled(T::AccountId, T::AccountId, u64, u16),
    AlphaBurned(T::AccountId, T::AccountId, u64, u16),
    EvmKeyAssociated {
        netuid: u16,
        hotkey: T::AccountId,
        evm_key: H160,
        block_associated: u64,
    },
    CRV3WeightsRevealed(u16, T::AccountId),
    CommitRevealPeriodsSet(u16, u64),
    CommitRevealEnabled(u16, bool),
    // some variants omitted

}

The Event enum of this pallet

Variants

NetworkAdded(u16, u16)

a new network is added.

NetworkRemoved(u16)

a network is removed.

StakeAdded(T::AccountId, T::AccountId, u64, u64, u16, u64)

stake has been transferred from the a coldkey account onto the hotkey staking account.

StakeRemoved(T::AccountId, T::AccountId, u64, u64, u16, u64)

stake has been removed from the hotkey staking account onto the coldkey account.

AggregatedStakeAdded(T::AccountId, T::AccountId, u16, u64)

stake has been transferred from the coldkey account onto the hotkey staking account (at the end of the block)

FailedToAddAggregatedStake(T::AccountId, T::AccountId, u16, u64)

adding aggregated stake has failed

AggregatedLimitedStakeAdded(T::AccountId, T::AccountId, u16, u64, u64, bool)

limited stake has been transferred from the coldkey account onto the hotkey staking account (at the end of the block)

FailedToAddAggregatedLimitedStake(T::AccountId, T::AccountId, u16, u64, u64, bool)

adding limited aggregated stake has failed

AggregatedStakeRemoved(T::AccountId, T::AccountId, u16, u64)

stake has been removed from the hotkey staking account into the coldkey account (at the end of the block).

FailedToRemoveAggregatedStake(T::AccountId, T::AccountId, u16, u64)

removing aggregated stake has failed

AggregatedLimitedStakeRemoved(T::AccountId, T::AccountId, u16, u64, u64, bool)

aggregated limited stake has been removed from the hotkey staking account into the coldkey account (at the end of the block).

FailedToRemoveAggregatedLimitedStake(T::AccountId, T::AccountId, u16, u64, u64, bool)

removing limited aggregated stake has failed

AggregatedUnstakeAllSucceeded(T::AccountId, T::AccountId)

aggregated unstake_all operation has succeeded

AggregatedUnstakeAllFailed(T::AccountId, T::AccountId)

aggregated unstake_all operation has failed

AggregatedUnstakeAllAlphaSucceeded(T::AccountId, T::AccountId)

aggregated unstake_all_alpha operation has succeeded

AggregatedUnstakeAllAlphaFailed(T::AccountId, T::AccountId)

aggregated unstake_all_alpha operation has failed

StakeMoved(T::AccountId, T::AccountId, u16, T::AccountId, u16, u64)

stake has been moved from origin (hotkey, subnet ID) to destination (hotkey, subnet ID) of this amount (in TAO).

WeightsSet(u16, u16)

a caller successfully sets their weights on a subnetwork.

NeuronRegistered(u16, u16, T::AccountId)

a new neuron account has been registered to the chain.

BulkNeuronsRegistered(u16, u16)

multiple uids have been concurrently registered.

BulkBalancesSet(u16, u16)

FIXME: Not used yet

MaxAllowedUidsSet(u16, u16)

max allowed uids has been set for a subnetwork.

MaxWeightLimitSet(u16, u16)

the max weight limit has been set for a subnetwork.

DifficultySet(u16, u64)

the difficulty has been set for a subnet.

AdjustmentIntervalSet(u16, u16)

the adjustment interval is set for a subnet.

RegistrationPerIntervalSet(u16, u16)

registration per interval is set for a subnet.

MaxRegistrationsPerBlockSet(u16, u16)

we set max registrations per block.

ActivityCutoffSet(u16, u16)

an activity cutoff is set for a subnet.

RhoSet(u16, u16)

Rho value is set.

AlphaSigmoidSteepnessSet(u16, u16)

steepness of the sigmoid used to compute alpha values.

KappaSet(u16, u16)

Kappa is set for a subnet.

MinAllowedWeightSet(u16, u16)

minimum allowed weight is set for a subnet.

ValidatorPruneLenSet(u16, u64)

the validator pruning length has been set.

ScalingLawPowerSet(u16, u16)

the scaling law power has been set for a subnet.

WeightsSetRateLimitSet(u16, u64)

weights set rate limit has been set for a subnet.

ImmunityPeriodSet(u16, u16)

immunity period is set for a subnet.

BondsMovingAverageSet(u16, u64)

bonds moving average is set for a subnet.

BondsPenaltySet(u16, u16)

bonds penalty is set for a subnet.

BondsResetOnSet(u16, bool)

bonds reset is set for a subnet.

MaxAllowedValidatorsSet(u16, u16)

setting the max number of allowed validators on a subnet.

AxonServed(u16, T::AccountId)

the axon server information is added to the network.

PrometheusServed(u16, T::AccountId)

the prometheus server information is added to the network.

DelegateAdded(T::AccountId, T::AccountId, u16)

a hotkey has become a delegate.

DefaultTakeSet(u16)

the default take is set.

WeightsVersionKeySet(u16, u64)

weights version key is set for a network.

MinDifficultySet(u16, u64)

setting min difficulty on a network.

MaxDifficultySet(u16, u64)

setting max difficulty on a network.

ServingRateLimitSet(u16, u64)

setting the prometheus serving rate limit.

BurnSet(u16, u64)

setting burn on a network.

MaxBurnSet(u16, u64)

setting max burn on a network.

MinBurnSet(u16, u64)

setting min burn on a network.

TxRateLimitSet(u64)

setting the transaction rate limit.

TxDelegateTakeRateLimitSet(u64)

setting the delegate take transaction rate limit.

TxChildKeyTakeRateLimitSet(u64)

setting the childkey take transaction rate limit.

MinChildKeyTakeSet(u16)

minimum childkey take set

MaxChildKeyTakeSet(u16)

maximum childkey take set

ChildKeyTakeSet(T::AccountId, u16)

childkey take set

Sudid(DispatchResult)

a sudo call is done.

RegistrationAllowed(u16, bool)

registration is allowed/disallowed for a subnet.

PowRegistrationAllowed(u16, bool)

POW registration is allowed/disallowed for a subnet.

TempoSet(u16, u16)

setting tempo on a network

RAORecycledForRegistrationSet(u16, u64)

setting the RAO recycled for registration.

StakeThresholdSet(u64)

min stake is set for validators to set weights.

SenateRequiredStakePercentSet(u64)

setting the minimum required stake amount for senate registration.

AdjustmentAlphaSet(u16, u64)

setting the adjustment alpha on a subnet.

Faucet(T::AccountId, u64)

the faucet it called on the test net.

SubnetOwnerCutSet(u16)

the subnet owner cut is set.

NetworkRateLimitSet(u64)

the network creation rate limit is set.

NetworkImmunityPeriodSet(u64)

the network immunity period is set.

NetworkMinLockCostSet(u64)

the network minimum locking cost is set.

NetworkLockCostReductionIntervalSet(u64)

the maximum number of subnets is set the lock cost reduction is set

TakeDecreased(T::AccountId, T::AccountId, u16)

the take for a delegate is decreased.

TakeIncreased(T::AccountId, T::AccountId, u16)

the take for a delegate is increased.

HotkeySwapped

the hotkey is swapped

Fields

coldkey: T::AccountId

the account ID of coldkey

old_hotkey: T::AccountId

the account ID of old hotkey

new_hotkey: T::AccountId

the account ID of new hotkey

MaxDelegateTakeSet(u16)

maximum delegate take is set by sudo/admin transaction

MinDelegateTakeSet(u16)

minimum delegate take is set by sudo/admin transaction

SenateAdjusted

a member of the senate is adjusted

Fields

old_member: Option<T::AccountId>

the account ID of the old senate member, if any

new_member: T::AccountId

the account ID of the new senate member

ColdkeySwapped

A coldkey has been swapped

Fields

old_coldkey: T::AccountId

the account ID of old coldkey

new_coldkey: T::AccountId

the account ID of new coldkey

swap_cost: u64

the swap cost

AllBalanceUnstakedAndTransferredToNewColdkey

All balance of a hotkey has been unstaked and transferred to a new coldkey

Fields

current_coldkey: T::AccountId

The account ID of the current coldkey

new_coldkey: T::AccountId

The account ID of the new coldkey

total_balance: <<T as Config>::Currency as Inspect<<T as Config>::AccountId>>::Balance

The total balance of the hotkey

ColdkeySwapScheduled

A coldkey swap has been scheduled

Fields

old_coldkey: T::AccountId

The account ID of the old coldkey

new_coldkey: T::AccountId

The account ID of the new coldkey

execution_block: BlockNumberFor<T>

The arbitration block for the coldkey swap

swap_cost: u64

The swap cost

ArbitrationPeriodExtended

The arbitration period has been extended

Fields

coldkey: T::AccountId

The account ID of the coldkey

SetChildrenScheduled(T::AccountId, u16, u64, Vec<(u64, T::AccountId)>)

Setting of children of a hotkey have been scheduled

SetChildren(T::AccountId, u16, Vec<(u64, T::AccountId)>)

The children of a hotkey have been set

ChainIdentitySet(T::AccountId)

The identity of a coldkey has been set

SubnetIdentitySet(u16)

The identity of a subnet has been set

SubnetIdentityRemoved(u16)

The identity of a subnet has been removed

DissolveNetworkScheduled

A dissolve network extrinsic scheduled.

Fields

account: T::AccountId

The account ID schedule the dissolve network extrisnic

netuid: u16

network ID will be dissolved

execution_block: BlockNumberFor<T>

extrinsic execution block number

ColdkeySwapScheduleDurationSet(BlockNumberFor<T>)

The duration of schedule coldkey swap has been set

DissolveNetworkScheduleDurationSet(BlockNumberFor<T>)

The duration of dissolve network has been set

CRV3WeightsCommitted(T::AccountId, u16, H256)

Commit-reveal v3 weights have been successfully committed.

who: The account ID of the user committing the weights.

netuid: The network identifier.

commit_hash: The hash representing the committed weights.

WeightsCommitted(T::AccountId, u16, H256)

Weights have been successfully committed.

who: The account ID of the user committing the weights.

netuid: The network identifier.

commit_hash: The hash representing the committed weights.

WeightsRevealed(T::AccountId, u16, H256)

Weights have been successfully revealed.

who: The account ID of the user revealing the weights.

netuid: The network identifier.

commit_hash: The hash of the revealed weights.

WeightsBatchRevealed(T::AccountId, u16, Vec<H256>)

Weights have been successfully batch revealed.

who: The account ID of the user revealing the weights.

netuid: The network identifier.

revealed_hashes: A vector of hashes representing each revealed weight set.

BatchWeightsCompleted(Vec<Compact<u16>>, T::AccountId)

A batch of weights (or commits) have been force-set.

netuids: The netuids these weights were successfully set/committed for.

who: The hotkey that set this batch.

BatchCompletedWithErrors()

A batch extrinsic completed but with some errors.

BatchWeightItemFailed(DispatchError)

A weight set among a batch of weights failed.

error: The dispatch error emitted by the failed item.

StakeTransferred(T::AccountId, T::AccountId, T::AccountId, u16, u16, u64)

Stake has been transferred from one coldkey to another on the same subnet. Parameters: (origin_coldkey, destination_coldkey, hotkey, origin_netuid, destination_netuid, amount)

StakeSwapped(T::AccountId, T::AccountId, u16, u16, u64)

Stake has been swapped from one subnet to another for the same coldkey-hotkey pair.

Parameters: (coldkey, hotkey, origin_netuid, destination_netuid, amount)

TransferToggle(u16, bool)

Event called when transfer is toggled on a subnet.

Parameters: (netuid, bool)

SubnetOwnerHotkeySet(u16, T::AccountId)

The owner hotkey for a subnet has been set.

Parameters: (netuid, new_hotkey)

FirstEmissionBlockNumberSet(u16, u64)

FirstEmissionBlockNumber is set via start call extrinsic

Parameters: netuid block number

AlphaRecycled(T::AccountId, T::AccountId, u64, u16)

Alpha has been recycled, reducing AlphaOut on a subnet.

Parameters: (coldkey, hotkey, amount, subnet_id)

AlphaBurned(T::AccountId, T::AccountId, u64, u16)

Alpha have been burned without reducing AlphaOut.

Parameters: (coldkey, hotkey, amount, subnet_id)

EvmKeyAssociated

An EVM key has been associated with a hotkey.

Fields

netuid: u16

The subnet that the hotkey belongs to.

hotkey: T::AccountId

The hotkey associated with the EVM key.

evm_key: H160

The EVM key being associated with the hotkey.

block_associated: u64

The block where the association happened.

CRV3WeightsRevealed(u16, T::AccountId)

CRV3 Weights have been successfully revealed.

netuid: The network identifier.

who: The account ID of the user revealing the weights.

CommitRevealPeriodsSet(u16, u64)

Commit-Reveal periods has been successfully set.

netuid: The network identifier.

periods: The number of epochs before the reveal.

CommitRevealEnabled(u16, bool)

Commit-Reveal has been successfully toggled.

netuid: The network identifier.

Enabled: Is Commit-Reveal enabled.