pallet_subtensor::pallet

Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

Active

— MAP ( netuid ) –> active

ActivityCutoff

— MAP ( netuid ) –> activity_cutoff

AdjustmentAlpha

— MAP ( netuid ) –> adjustment_alpha

AdjustmentInterval

— MAP ( netuid ) –> adjustment_interval

Alpha

Warning: Doc-Only

AlphaDividendsPerSubnet

Warning: Doc-Only

AlphaSigmoidSteepness

— MAP ( netuid ) –> AlphaSigmoidSteepness

AlphaValues

MAP ( netuid ) –> (alpha_low, alpha_high)

AssociatedEvmAddress

============================= ==== EVM related storage ====

Axons

— MAP ( netuid, hotkey ) –> axon_info

BlockAtRegistration

— DMAP ( netuid, uid ) –> block_at_registration

BlockEmission

================== ==== Coinbase ====

BlocksSinceLastStep

— MAP ( netuid ) –> blocks_since_last_step

Bonds

— DMAP ( netuid, uid ) –> bonds

BondsMovingAverage

— MAP ( netuid ) –> bonds_moving_average

BondsPenalty

— MAP ( netuid ) –> bonds_penalty

BondsResetOn

— MAP ( netuid ) –> bonds_reset

Burn

— MAP ( netuid ) –> Burn

BurnRegistrationsThisInterval

— MAP ( netuid ) –> burn_registrations_this_interval

CRV3WeightCommits

— MAP (netuid, commit_epoch) –> VecDeque<(who, serialized_compressed_commit, reveal_round)> | Stores a queue of v3 commits for an account on a given netuid.

ChildKeys

DMAP ( parent, netuid ) –> Vec<(proportion,child)>

ChildkeyTake

DMAP ( hot, netuid ) –> take | Returns the hotkey childkey take for a specific subnet

ColdkeySwapRescheduleDuration

Warning: Doc-Only

ColdkeySwapScheduleDuration

Warning: Doc-Only

ColdkeySwapScheduled

Warning: Doc-Only

CommitRevealWeightsEnabled

— MAP ( netuid ) –> commit reveal v2 weights are enabled

Consensus

— MAP ( netuid ) –> consensus

Delegates

MAP ( hot ) –> take | Returns the hotkey delegation take. And signals that this key is open for delegation.

Difficulty

— MAP ( netuid ) –> Difficulty

DissolveNetworkScheduleDuration

Warning: Doc-Only

Dividends

— MAP ( netuid ) –> dividends

EMAPriceHalvingBlocks

— MAP ( netuid ) –> Halving time of average moving price.

Emission

— MAP ( netuid ) –> emission

FirstEmissionBlockNumber

============================ ==== Subnet Parameters =====

HasMigrationRun

================== ==== Genesis =====

Identities

Warning: Doc-Only

IdentitiesV2

Warning: Doc-Only

ImmunityPeriod

— MAP ( netuid ) –> immunity_period

Incentive

— MAP ( netuid ) –> incentive

IsNetworkMember

— DMAP ( hotkey, netuid ) –> bool

Kappa

— MAP ( netuid ) –> Kappa

Keys

— DMAP ( netuid, uid ) –> hotkey

LargestLocked

Warning: Doc-Only

LastAdjustmentBlock

— MAP ( netuid ) –> Block at last adjustment.

LastColdkeyHotkeyStakeBlock

— Map (coldkey, hotkey) –> u64 the last block at which stake was added/removed.

LastHotkeyEmissionOnNetuid

— DMap ( hot, netuid ) –> emission | last hotkey emission on network.

LastMechansimStepBlock

— MAP ( netuid ) –> last_mechanism_step_block

LastRateLimitedBlock

============================ ==== Rate Limiting =====

LastTxBlock

— MAP ( key ) –> last_block

LastTxBlockChildKeyTake

— MAP ( key ) –> last_tx_block_childkey_take

LastTxBlockDelegateTake

— MAP ( key ) –> last_tx_block_delegate_take

LastUpdate

— MAP ( netuid ) –> last_update

LiquidAlphaOn

— MAP ( netuid ) –> Whether or not Liquid Alpha is enabled

LoadedEmission

— MAP ( netuid ) –> (hotkey, se, ve)

MaxAllowedUids

— MAP ( netuid ) –> max_allowed_uids

MaxAllowedValidators

— MAP ( netuid ) –> max_allowed_validators

MaxBurn

— MAP ( netuid ) –> MaxBurn

MaxChildkeyTake

— ITEM ( default_childkey_take )

MaxDelegateTake

— ITEM ( default_delegate_take )

MaxDifficulty

— MAP ( netuid ) –> MaxDifficulty

MaxRegistrationsPerBlock

— ITEM( global_max_registrations_per_block )

MaxWeightsLimit

— MAP ( netuid ) –> max_weight_limit

MinActivityCutoff

Warning: Doc-Only

MinAllowedWeights

— MAP ( netuid ) –> min_allowed_weights

MinBurn

— MAP ( netuid ) –> MinBurn

MinChildkeyTake

— ITEM ( min_childkey_take )

MinDelegateTake

— ITEM ( min_delegate_take )

MinDifficulty

— MAP ( netuid ) –> MinDifficulty

NetworkImmunityPeriod

ITEM( network_immunity_period )

NetworkLastLockCost

ITEM( last_network_lock_cost )

NetworkLastRegistered

ITEM( network_last_registered_block )

NetworkLockReductionInterval

ITEM( network_lock_reduction_interval )

NetworkMinLockCost

ITEM( min_network_lock_cost )

NetworkModality

— MAP ( netuid ) –> modality TEXT: 0, IMAGE: 1, TENSOR: 2

NetworkPowRegistrationAllowed

— MAP ( netuid ) –> network_pow_allowed

NetworkRateLimit

ITEM( network_rate_limit )

NetworkRegisteredAt

— MAP ( netuid ) –> block_created

NetworkRegistrationAllowed

— MAP ( netuid ) –> network_registration_allowed

NetworksAdded

— MAP ( netuid ) –> network_is_added

NeuronCertificates

— MAP ( netuid, hotkey ) –> certificate

NextStakeJobId

Ensures unique IDs for StakeJobs storage map

NominatorMinRequiredStake

Warning: Doc-Only

OwnedHotkeys

Warning: Doc-Only

Owner

MAP ( hot ) –> cold | Returns the controlling coldkey for a hotkey.

POWRegistrationsThisInterval

— MAP ( netuid ) –> pow_registrations_this_interval

ParentKeys

DMAP ( child, netuid ) –> Vec<(proportion,parent)>

PendingAlphaSwapped

— MAP ( netuid ) –> pending_alpha_swapped

PendingChildKeys

DMAP ( netuid, parent ) –> (Vec<(proportion,child)>, cool_down_block)

PendingEmission

— MAP ( netuid ) –> pending_emission

PendingOwnerCut

— MAP ( netuid ) –> pending_owner_cut

PendingRootDivs

— MAP ( netuid ) –> pending_root_emission

Prometheus

— MAP ( netuid, hotkey ) –> prometheus_info

PruningScores

— MAP ( netuid ) –> pruning_scores

RAORecycledForRegistration

— MAP ( netuid ) –> global_RAO_recycled_for_registration

Rank

— MAP ( netuid ) –> rank

RegistrationsThisBlock

— MAP ( netuid ) –> Registrations of this Block.

RegistrationsThisInterval

— MAP ( netuid ) –> registrations_this_interval

RevealPeriodEpochs

— Map (netuid) –> Number of epochs allowed for commit reveal periods

Rho

— MAP ( netuid ) –> Rho

ScalingLawPower

— MAP ( netuid ) –> scaling_law_power

SenateRequiredStakePercentage

Warning: Doc-Only

ServingRateLimit

— MAP ( netuid ) –> serving_rate_limit

StakeJobs

Warning: Doc-Only

StakeThreshold

ITEM( weights_min_stake )

StakeWeight

======================================= ==== Subnetwork Consensus Storage ====

StakingHotkeys

Warning: Doc-Only

SubnetAlphaIn

Warning: Doc-Only

SubnetAlphaInEmission

Warning: Doc-Only

SubnetAlphaOut

Warning: Doc-Only

SubnetAlphaOutEmission

Warning: Doc-Only

SubnetIdentities

Warning: Doc-Only

SubnetIdentitiesV2

Warning: Doc-Only

SubnetLocked

Warning: Doc-Only

SubnetMechanism

— MAP ( netuid ) –> subnet mechanism

SubnetMovingAlpha

Warning: Doc-Only

SubnetMovingPrice

Warning: Doc-Only

SubnetOwner

— MAP ( netuid ) –> subnet_owner

SubnetOwnerCut

ITEM( subnet_owner_cut )

SubnetOwnerHotkey

— MAP ( netuid ) –> subnet_owner_hotkey

SubnetTAO

Warning: Doc-Only

SubnetTaoInEmission

Warning: Doc-Only

SubnetVolume

Warning: Doc-Only

SubnetworkN

— MAP ( netuid ) –> subnetwork_n (Number of UIDs in the network).

SubtokenEnabled

— MAP ( netuid ) –> If subtoken trading enabled

TaoDividendsPerSubnet

Warning: Doc-Only

TaoWeight

============================ ==== Staking Variables ====

TargetRegistrationsPerInterval

— MAP ( netuid ) –> target_registrations_this_interval

Tempo

================= ==== Tempos =====

TokenSymbol

Warning: Doc-Only

TotalHotkeyAlpha

Warning: Doc-Only

TotalHotkeyAlphaLastEpoch

Warning: Doc-Only

TotalHotkeyShares

DMAP ( hot, netuid ) –> total_alpha_shares | Returns the number of alpha shares for a hotkey on a subnet.

TotalIssuance

========================== ==== Staking Counters ====

TotalNetworks

— ITEM( total_number_of_existing_networks )

TotalStake

Warning: Doc-Only

TransactionKeyLastBlock

================================= ==== Axon / Promo Endpoints =====

TransferToggle

============================ ==== Subnet Locks =====

Trust

— MAP ( netuid ) –> trust

TxChildkeyTakeRateLimit

— ITEM ( tx_childkey_take_rate_limit )

TxDelegateTakeRateLimit

— ITEM ( tx_delegate_take_rate_limit )

TxRateLimit

— ITEM ( tx_rate_limit )

Uids

— DMAP ( netuid, hotkey ) –> uid

UsedWork

============================ ==== Global Parameters =====

ValidatorPermit

— MAP ( netuid ) –> validator_permit

ValidatorPruneLen

— MAP ( netuid ) –> validator_prune_len

ValidatorTrust

— MAP ( netuid ) –> validator_trust

WeightCommits

— MAP (netuid, who) –> VecDeque<(hash, commit_block, first_reveal_block, last_reveal_block)> | Stores a queue of commits for an account on a given netuid.

Weights

— DMAP ( netuid, uid ) –> weights

WeightsSetRateLimit

— MAP ( netuid ) –> weights_set_rate_limit

WeightsVersionKey

— MAP ( netuid ) –> weights_version_key

WeightsVersionKeyRateLimit

ITEM( weights_version_key_rate_limit ) — Rate limit in tempos.

Yuma3On

— MAP ( netuid ) –> Whether or not Yuma3 is enabled