pallet_subtensor::pallet

Trait Config

pub trait Config: Config + Config {

    type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + From<Call<Self>> + IsType<<Self as Config>::RuntimeCall> + From<Call<Self>>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type SudoRuntimeCall: Parameter + UnfilteredDispatchable<RuntimeOrigin = Self::RuntimeOrigin> + GetDispatchInfo;
    type CouncilOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type Currency: Balanced<Self::AccountId, Balance = u64> + Mutate<Self::AccountId>;
    type SenateMembers: MemberManagement<Self::AccountId>;
    type TriumvirateInterface: CollectiveInterface<Self::AccountId, Self::Hash, u32>;
    type Scheduler: ScheduleAnon<BlockNumberFor<Self>, LocalCallOf<Self>, PalletsOriginOf<Self>, Hasher = Self::Hashing>;
    type Preimages: QueryPreimage<H = Self::Hashing> + StorePreimage;
    type InitialIssuance: Get<u64>;
    type InitialMinAllowedWeights: Get<u16>;
    type InitialEmissionValue: Get<u16>;
    type InitialMaxWeightsLimit: Get<u16>;
    type InitialTempo: Get<u16>;
    type InitialDifficulty: Get<u64>;
    type InitialMaxDifficulty: Get<u64>;
    type InitialMinDifficulty: Get<u64>;
    type InitialRAORecycledForRegistration: Get<u64>;
    type InitialBurn: Get<u64>;
    type InitialMaxBurn: Get<u64>;
    type InitialMinBurn: Get<u64>;
    type InitialAdjustmentInterval: Get<u16>;
    type InitialBondsMovingAverage: Get<u64>;
    type InitialBondsPenalty: Get<u16>;
    type InitialBondsResetOn: Get<bool>;
    type InitialTargetRegistrationsPerInterval: Get<u16>;
    type InitialRho: Get<u16>;
    type InitialAlphaSigmoidSteepness: Get<u16>;
    type InitialKappa: Get<u16>;
    type InitialMaxAllowedUids: Get<u16>;
    type InitialValidatorPruneLen: Get<u64>;
    type InitialScalingLawPower: Get<u16>;
    type InitialImmunityPeriod: Get<u16>;
    type InitialActivityCutoff: Get<u16>;
    type InitialMaxRegistrationsPerBlock: Get<u16>;
    type InitialPruningScore: Get<u16>;
    type InitialMaxAllowedValidators: Get<u16>;
    type InitialDefaultDelegateTake: Get<u16>;
    type InitialMinDelegateTake: Get<u16>;
    type InitialDefaultChildKeyTake: Get<u16>;
    type InitialMinChildKeyTake: Get<u16>;
    type InitialMaxChildKeyTake: Get<u16>;
    type InitialWeightsVersionKey: Get<u64>;
    type InitialServingRateLimit: Get<u64>;
    type InitialTxRateLimit: Get<u64>;
    type InitialTxDelegateTakeRateLimit: Get<u64>;
    type InitialTxChildKeyTakeRateLimit: Get<u64>;
    type InitialSenateRequiredStakePercentage: Get<u64>;
    type InitialAdjustmentAlpha: Get<u64>;
    type InitialNetworkImmunityPeriod: Get<u64>;
    type InitialNetworkMinAllowedUids: Get<u16>;
    type InitialNetworkMinLockCost: Get<u64>;
    type InitialSubnetOwnerCut: Get<u16>;
    type InitialNetworkLockReductionInterval: Get<u64>;
    type InitialNetworkRateLimit: Get<u64>;
    type KeySwapCost: Get<u64>;
    type AlphaHigh: Get<u16>;
    type AlphaLow: Get<u16>;
    type LiquidAlphaOn: Get<bool>;
    type Yuma3On: Get<bool>;
    type InitialColdkeySwapScheduleDuration: Get<BlockNumberFor<Self>>;
    type InitialColdkeySwapRescheduleDuration: Get<BlockNumberFor<Self>>;
    type InitialDissolveNetworkScheduleDuration: Get<BlockNumberFor<Self>>;
    type InitialTaoWeight: Get<u64>;
    type InitialEmaPriceHalvingPeriod: Get<u64>;
    type DurationOfStartCall: Get<u64>;

}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type RuntimeCall: Parameter + Dispatchable<RuntimeOrigin = Self::RuntimeOrigin> + From<Call<Self>> + IsType<<Self as Config>::RuntimeCall> + From<Call<Self>>

call type

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type SudoRuntimeCall: Parameter + UnfilteredDispatchable<RuntimeOrigin = Self::RuntimeOrigin> + GetDispatchInfo

A sudo-able call.

type CouncilOrigin: EnsureOrigin<Self::RuntimeOrigin>

Origin checking for council majority

type Currency: Balanced<Self::AccountId, Balance = u64> + Mutate<Self::AccountId>

Currency type that will be used to place deposits on neurons

type SenateMembers: MemberManagement<Self::AccountId>

Senate members with members management functions.

type TriumvirateInterface: CollectiveInterface<Self::AccountId, Self::Hash, u32>

Interface to allow other pallets to control who can register identities

type Scheduler: ScheduleAnon<BlockNumberFor<Self>, LocalCallOf<Self>, PalletsOriginOf<Self>, Hasher = Self::Hashing>

The scheduler type used for scheduling delayed calls.

type Preimages: QueryPreimage<H = Self::Hashing> + StorePreimage

the preimage to store the call data.

type InitialIssuance: Get<u64>

================================= ==== Initial Value Constants ====

Initial currency issuance.

type InitialMinAllowedWeights: Get<u16>

Initial min allowed weights setting.

type InitialEmissionValue: Get<u16>

Initial Emission Ratio.

type InitialMaxWeightsLimit: Get<u16>

Initial max weight limit.

type InitialTempo: Get<u16>

Tempo for each network.

type InitialDifficulty: Get<u64>

Initial Difficulty.

type InitialMaxDifficulty: Get<u64>

Initial Max Difficulty.

type InitialMinDifficulty: Get<u64>

Initial Min Difficulty.

type InitialRAORecycledForRegistration: Get<u64>

Initial RAO Recycled.

type InitialBurn: Get<u64>

Initial Burn.

type InitialMaxBurn: Get<u64>

Initial Max Burn.

type InitialMinBurn: Get<u64>

Initial Min Burn.

type InitialAdjustmentInterval: Get<u16>

Initial adjustment interval.

type InitialBondsMovingAverage: Get<u64>

Initial bonds moving average.

type InitialBondsPenalty: Get<u16>

Initial bonds penalty.

type InitialBondsResetOn: Get<bool>

Initial bonds reset.

type InitialTargetRegistrationsPerInterval: Get<u16>

Initial target registrations per interval.

type InitialRho: Get<u16>

Rho constant.

type InitialAlphaSigmoidSteepness: Get<u16>

AlphaSigmoidSteepness constant.

type InitialKappa: Get<u16>

Kappa constant.

type InitialMaxAllowedUids: Get<u16>

Max UID constant.

type InitialValidatorPruneLen: Get<u64>

Initial validator context pruning length.

type InitialScalingLawPower: Get<u16>

Initial scaling law power.

type InitialImmunityPeriod: Get<u16>

Immunity Period Constant.

type InitialActivityCutoff: Get<u16>

Activity constant.

type InitialMaxRegistrationsPerBlock: Get<u16>

Initial max registrations per block.

type InitialPruningScore: Get<u16>

Initial pruning score for each neuron.

type InitialMaxAllowedValidators: Get<u16>

Initial maximum allowed validators per network.

type InitialDefaultDelegateTake: Get<u16>

Initial default delegation take.

type InitialMinDelegateTake: Get<u16>

Initial minimum delegation take.

type InitialDefaultChildKeyTake: Get<u16>

Initial default childkey take.

type InitialMinChildKeyTake: Get<u16>

Initial minimum childkey take.

type InitialMaxChildKeyTake: Get<u16>

Initial maximum childkey take.

type InitialWeightsVersionKey: Get<u64>

Initial weights version key.

type InitialServingRateLimit: Get<u64>

Initial serving rate limit.

type InitialTxRateLimit: Get<u64>

Initial transaction rate limit.

type InitialTxDelegateTakeRateLimit: Get<u64>

Initial delegate take transaction rate limit.

type InitialTxChildKeyTakeRateLimit: Get<u64>

Initial childkey take transaction rate limit.

type InitialSenateRequiredStakePercentage: Get<u64>

Initial percentage of total stake required to join senate.

type InitialAdjustmentAlpha: Get<u64>

Initial adjustment alpha on burn and pow.

type InitialNetworkImmunityPeriod: Get<u64>

Initial network immunity period

type InitialNetworkMinAllowedUids: Get<u16>

Initial minimum allowed network UIDs

type InitialNetworkMinLockCost: Get<u64>

Initial network minimum burn cost

type InitialSubnetOwnerCut: Get<u16>

Initial network subnet cut.

type InitialNetworkLockReductionInterval: Get<u64>

Initial lock reduction interval.

type InitialNetworkRateLimit: Get<u64>

Initial network creation rate limit

type KeySwapCost: Get<u64>

Cost of swapping a hotkey.

type AlphaHigh: Get<u16>

The upper bound for the alpha parameter. Used for Liquid Alpha.

type AlphaLow: Get<u16>

The lower bound for the alpha parameter. Used for Liquid Alpha.

type LiquidAlphaOn: Get<bool>

A flag to indicate if Liquid Alpha is enabled.

type Yuma3On: Get<bool>

A flag to indicate if Yuma3 is enabled.

type InitialColdkeySwapScheduleDuration: Get<BlockNumberFor<Self>>

Coldkey swap schedule duartion.

type InitialColdkeySwapRescheduleDuration: Get<BlockNumberFor<Self>>

Coldkey swap reschedule duration.

type InitialDissolveNetworkScheduleDuration: Get<BlockNumberFor<Self>>

Dissolve network schedule duration

type InitialTaoWeight: Get<u64>

Initial TAO weight.

type InitialEmaPriceHalvingPeriod: Get<u64>

Initial EMA price halving period

type DurationOfStartCall: Get<u64>

Block number after a new subnet accept the start call extrinsic.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.