pallet_subtensor

Trait MemberManagement

pub trait MemberManagement<AccountId> {
    // Required methods
    fn add_member(account: &AccountId) -> DispatchResultWithPostInfo;
    fn remove_member(account: &AccountId) -> DispatchResultWithPostInfo;
    fn swap_member(
        remove: &AccountId,
        add: &AccountId,
    ) -> DispatchResultWithPostInfo;
    fn members() -> Vec<AccountId>;
    fn is_member(account: &AccountId) -> bool;
    fn max_members() -> u32;
}

Trait for managing a membership pallet instance in the runtime

Required Methods

fn add_member(account: &AccountId) -> DispatchResultWithPostInfo

Add member

fn remove_member(account: &AccountId) -> DispatchResultWithPostInfo

Remove a member

fn swap_member( remove: &AccountId, add: &AccountId, ) -> DispatchResultWithPostInfo

Swap member

fn members() -> Vec<AccountId>

Get all members

fn is_member(account: &AccountId) -> bool

Check if an account is apart of the set

fn max_members() -> u32

Get our maximum member count

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.